### 2025-09-02 73543c5
Fixed ORCID o Viktor Makarichev
