### 2025-09-04 427c48a
Reduced the size of the paper
