+++ paper-joss/paper.bib
@@ -269,3 +269,10 @@ doi = {10.1137/S1064827594276424},
   year = {2025},
   url = {https://github.com/datagrok-ai/public/tree/master/packages/DiffStudio}
 }
+
+@software{diffgrokperformance,
+  author = {{Datagrok Inc}},
+  title = {Diff-Grok performance},
+  year = {2025},
+  url = {https://github.com/datagrok-ai/diff-grok?tab=readme-ov-file#performance}
+}
