+++ paper-joss/reduced-paper.md
@@ -1,0 +1,237 @@
+---
+title: 'Diff Studio: Ecosystem for Interactive Modeling by Ordinary Differential Equations'
+tags:
+  - Differential equations
+  - JavaScript
+authors:
+  - name: Viktor Makarichev   
+    orcid: 0000-0003-1481-9132
+    affiliation: "1" 
+    corresponding: true 
+  - name: Larisa Bankurova
+    affiliation: "1"
+  - name: Gennadii Zakharov
+    affiliation: "1,3"
+  - name: Leonid Stolbov
+    affiliation: "1"
+  - name: Steven Mehrman
+    affiliation: "2"
+  - name: Dan Skatov
+    affiliation: "2"
+  - name: Jeffrey Cohen
+    affiliation: "2"
+  - name: Paul Sass
+    affiliation: "2"
+  - name: Davit Rizhinashvili 
+    affiliation: "1"
+  - name: Andrew Skalkin
+    affiliation: "1"
+affiliations:
+ - name: Datagrok Inc, USA
+   index: 1
+ - name: Johnson & Johnson Inc, USA
+   index: 2
+ - name: Wellcome Sanger Institute, UK
+   index: 3
+date: 04 September 2025
+bibliography: paper.bib
+---
+
+# Summary
+
+Ordinary differential equations (ODEs) are very important in modeling complex systems and phenomena. Their applications range from pharmacology and drug manufacturing 
+to financial modeling and environmental studies.
+
+**Diff Studio** is a high-performance TypeScript application 
+for solving initial value problems for ODEs directly within web browsers. It enables non-programmers to define models via a domain-specific language with real-time visualization, syntax support, error detection, and one-click generation of standalone ODE applications.
+
+Diff Studio consists of two components. 
+The **Diff Grok library** (<https://github.com/datagrok-ai/diff-grok>)
+implements numerical methods and formula parsing tools. 
+The **Diff Studio application** (<https://datagrok.ai/help/compute/diff-studio>)
+integrates Diff Grok tools with **Datagrok** (<https://datagrok.ai>),  a scientific computing platform free for personal and academic use.
+
+Diff Studio constitutes an ecosystem for rapid development of ODE-based applications that ensure reproducibility and accessibility through a shared model repository.
+
+
+# Statement of need
+
+Scientific modeling of complex processes and phenomena uses ordinary differential equations. They are widely applied in diverse fields, including physical processes [@chicone2006ordinary], 
+biochemical kinetics [@ingalls2013mathematical], drug delivery systems [@mircioiu2019mathematical], cloud computing [@jafarnejad2019applying], and population dynamics [@hastings2013population].
+
+Analytic methods providing exact solutions
+can be applied only to a limited class of ODEs. 
+The use of analytic
+solutions often proves impractical due to their complexity [@hairer2008solving1].
+Numerical methods computing approximate solutions
+are often preferred.
+
+Many methods for solving ODEs have been recently developed 
+[@hairer2008solving1; @hairer2002solving2]. 
+These methods have been implemented in various software tools, 
+including libraries and packages for programming languages and
+scientific computing environments. 
+Notable examples include 
+SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials], 
+Julia Differential Equations package [@rackauckas2017differentialequations], 
+SciPy [@2020SciPyNMeth], 
+Maple [@maple2025], 
+Mathematica [@Mathematica2024],
+Matlab [@MATLAB], 
+and deSolve [@soetaert2010solving].
+
+Though most existing tools require significant technical expertise,
+shifting emphasis from applied research to software development. 
+The goal of this project is to develop an ecosystem 
[…]
