### 2025-09-04 9c07785
Reduced the number of words of the paper
