+++ paper-joss/reduced-paper.md
@@ -42,29 +42,18 @@ bibliography: paper.bib
 Ordinary differential equations (ODEs) are very important in modeling complex systems and phenomena. Their applications range from pharmacology and drug manufacturing 
 to financial modeling and environmental studies.
 
-**Diff Studio** is a high-performance TypeScript application 
-for solving initial value problems for ODEs directly within web browsers. It enables non-programmers to define models via a domain-specific language with real-time visualization, syntax support, error detection, and one-click generation of standalone ODE applications.
+**Diff Studio** is a high-performance TypeScript application for solving initial value problems (IVPs) for ODEs directly within web browsers. It consists of two components. The **Diff Grok library** implements numerical methods and formula parsing tools. The **Diff Studio application** integrates Diff Grok tools with **Datagrok**, a scientific computing platform free for personal and academic use.
 
-Diff Studio consists of two components. 
-The **Diff Grok library** (<https://github.com/datagrok-ai/diff-grok>)
-implements numerical methods and formula parsing tools. 
-The **Diff Studio application** (<https://datagrok.ai/help/compute/diff-studio>)
-integrates Diff Grok tools with **Datagrok** (<https://datagrok.ai>),  a scientific computing platform free for personal and academic use.
-
-Diff Studio constitutes an ecosystem for rapid development of ODE-based applications that ensure reproducibility and accessibility through a shared model repository.
+Diff Studio provides an ecosystem for fast development of ODE-based applications with reproducible and accessible models.
 
 
 # Statement of need
 
-Scientific modeling of complex processes and phenomena uses ordinary differential equations. They are widely applied in diverse fields, including physical processes [@chicone2006ordinary], 
+Scientific modeling of complex processes and phenomena uses ODEs. They are widely applied in diverse fields, including physical processes [@chicone2006ordinary], 
 biochemical kinetics [@ingalls2013mathematical], drug delivery systems [@mircioiu2019mathematical], cloud computing [@jafarnejad2019applying], and population dynamics [@hastings2013population].
 
-Analytic methods providing exact solutions
-can be applied only to a limited class of ODEs. 
-The use of analytic
-solutions often proves impractical due to their complexity [@hairer2008solving1].
-Numerical methods computing approximate solutions
-are often preferred.
+Analytic methods providing exact solutions can be applied only to a limited class of ODEs. 
+The use of analytic solutions often proves impractical due to their complexity [@hairer2008solving1]. Numerical methods computing approximate solutions are often preferred.
 
 Many methods for solving ODEs have been recently developed 
 [@hairer2008solving1; @hairer2002solving2]. 
@@ -80,20 +69,16 @@ Mathematica [@Mathematica2024],
 Matlab [@MATLAB], 
 and deSolve [@soetaert2010solving].
 
-Though most existing tools require significant technical expertise,
-shifting emphasis from applied research to software development. 
-The goal of this project is to develop an ecosystem 
-providing comprehensive capabilities for in-browser ODE modeling and analysis.
+Most tools require expertise, shifting focus from research to development. 
+The goal of this project is to develop an ecosystem providing a combination of a "no-code" approach with comprehensive capabilities for in-browser modeling and analysis.
 
 # The solution: Diff Studio
-
-We introduce **Diff Studio** - a solution that lets users define complex models using the special declarative notation and interactively explore model behavior in a web browser. It consists of two components: the **Diff Grok library** [@diffgrok] and the **Diff Studio package** [@diffstudio].
 
[…]
