### 2025-09-04 51e4b49
Fixed the formula in the paper
