### 2025-09-05 #8907 @danielskatz
@vmakarichev - this has passed the JOSS scope review.

So I'm going to demonstrate commands to check the paper length and generate the pdf, which you should use as you make changes to reduce the paper's length.

I'm also going to mark this as paused until you let me know that I should look at it again.
