### 2025-09-05 e3db9fe
Reduced the size of the main paper
