+++ paper-joss/paper.md
@@ -33,55 +33,27 @@ affiliations:
    index: 2
  - name: Wellcome Sanger Institute, UK
    index: 3
-date: 03 April 2025
+date: 04 September 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-Differential equations play a crucial role in modeling complex systems and phenomena, 
-with applications ranging from pharmacology and drug manufacturing 
+Ordinary differential equations (ODEs) are crucial in modeling complex systems and phenomena. Their applications range from pharmacology and drug manufacturing 
 to financial modeling and environmental studies.
 
-**Diff Studio** is a high-performance TypeScript application 
-for solving initial value problems in ordinary differential equations (ODEs)
-directly within web browsers. 
-It enables users without programming expertise to declaratively define models 
-using a simple domain-specific language 
-and provides real-time solution visualization,
-syntax highlighting, error detection, 
-and one-click creation of standalone applications directly from an ODE's system. 
+**Diff Studio** is a high-performance TypeScript application for solving initial value problems (IVPs) for ODEs directly within web browsers. It consists of two components. The **Diff Grok library** implements numerical methods and formula parsing tools. The **Diff Studio application** integrates Diff Grok tools with **Datagrok**, a scientific computing platform free for personal and academic use.
 
-Diff Studio consists of two components. 
-The **Diff Grok library** (<https://github.com/datagrok-ai/diff-grok>)
-implements numerical methods and formula parsing tools. 
-The **Diff Studio application** (<https://datagrok.ai/help/compute/diff-studio>)
-integrates Diff Grok tools with 
-**Datagrok** (<https://datagrok.ai>), 
-a next-generation scientific computing platform.
+Diff Studio provides an ecosystem for rapid development of ODE-based applications with reproducible and accessible models.
 
-Diff Studio creates an ecosystem
-for rapid development of scientific applications based on ODE models. 
-It democratizes research by providing a central repository for ODE models 
-and guaranteeing the reproducibility and accessibility of results.
 
 # Statement of need
 
-Scientific modeling of complex processes and phenomena relies heavily on
-ordinary differential equations. 
-Their applications span diverse fields,
-including physical processes [@chicone2006ordinary], 
-biochemical kinetics [@ingalls2013mathematical], 
-drug delivery systems [@mircioiu2019mathematical], 
-cloud computing [@jafarnejad2019applying], 
-and population dynamics [@hastings2013population].
+Scientific modeling of complex processes and phenomena uses ODEs. They are widely applied in diverse fields, including physical processes [@chicone2006ordinary], 
+biochemical kinetics [@ingalls2013mathematical], drug delivery systems [@mircioiu2019mathematical], cloud computing [@jafarnejad2019applying], and population dynamics [@hastings2013population].
 
-Analytic methods providing exact solutions
-can be applied only to a limited class of ODEs. 
-The use of analytic
-solutions often proves impractical due to their complexity [@hairer2008solving1].
-Numerical methods computing approximate solutions
-are often preferred.
+Analytic methods providing exact solutions can be applied only to a limited class of ODEs. 
+The use of analytic solutions often proves impractical due to their complexity [@hairer2008solving1]. Numerical methods computing approximate solutions are often preferred.
 
 Many methods for solving ODEs have been recently developed 
 [@hairer2008solving1; @hairer2002solving2]. 
@@ -97,51 +69,20 @@ Mathematica [@Mathematica2024],
 Matlab [@MATLAB], 
 and deSolve [@soetaert2010solving].
 
[…]
