### 2025-09-05 10a91e4
Removed the redundant text from the paper
