### 2025-09-16 #8907 @lockwo
Hi 👋 @ketch, @solanpaa, @avik-pal would one (or more) of you be willing to review this submission for JOSS?
