### 2025-09-27 #8907 @josemanuel22
@lockwo I won’t be able to start until October 10. After that, I’ll be available.
