### 2025-09-29 #9090 @editorialbot
[REVIEW]: Diff Studio: Ecosystem for Interactive Modeling by Ordinary Differential Equations
**Submitting author:** @vmakarichev (<a href="http://orcid.org/0000-0003-1481-9132">Victor Makarichev</a>)
**Repository:** https://github.com/datagrok-ai/diff-grok
**Branch with paper.md** (empty if default branch): gzakharov/DiffStudio-paper-JOSS
**Version:** v1.2.0
