### 2025-11-04 #9090 @lockwo
> Thank you for replying to my concerns. It seems lke most of the code the paper refers to is indeed not part of the submitted repository and according to the author, is proprietary, private and requires an NDA to access. Whether or not this falls under scope for JOSS should be addressed by [@lockwo](https://github.com/lockwo) in my opinion. In any case, most of the code comprising Diff Studio is not published and the only publicly accessible part is the Diff Grok library, which is minimal.

Thanks for raising this point, this concern is valid. In attempt to assuage some concerns I will note that during the pre-review the EiC (editor in chief) queried the scope (a process to bring together the editors to determine if a submission adheres to the requirements of JOSS) which did pass https://github.com/openjournals/joss-reviews/issues/8907#issuecomment-3258715493.
