### 2025-11-05 5a80bc8
GROK-19134: Diff Grok: Pipeline tests
Added tests for pipelines
