+++ src/tests/test-defs.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-len */
 import {mrt, ros3prw, ros34prw} from '../../index';
 
 export const methods = new Map([
@@ -8,3 +9,203 @@ export const methods = new Map([
 
 export const MAX_MAD = 0.1;
 export const TIMEOUT_MS = 10000;
+
+const BASIC_MODEL = `#name: M|M|2|2
+#description: Modelling the system with two servers & two waiting places (EXPLAINED)
+#equations:  
+  dp0/dt = -la * p0 + mu * p1
+  dp1/dt = la * p0 - (la + mu) * p1 + 2 * mu * p2
+  dp2/dt = la * p1 - (la + 2 * mu) * p2 + 2 * mu * p3
+  dp3/dt = la * p2 - (la + 2 * mu) * p3 + 2 * mu * p4
+
+#expressions:
+  la = 1 / arrival
+  mu = 1 / service
+  p4 = 1 - p0 - p1 - p2 - p3
+  busy = 1 - p0 - p1
+  queue = p3 + 2 * p4
+
+#argument: t
+  _t0 = 0     {min: 0;      max: 10;   caption: start;   category: Time; units: min}  [Initial time of simulation]
+  _t1 = 60    {min: 20;     max: 100;  caption: finish;  category: Time; units: min}  [Final time of simulation]
+  _h  = 1     {min: 0.01;   max: 0.1;  caption: step;    category: Time; units: min}  [Time step of simulation] 
+
+#inits:
+  p0 = 1   {min: 0; max: 1; category: Initial state; caption: empty}   [Probability that initially there are NO customers]
+  p1 = 0   {min: 0; max: 1; category: Initial state; caption: single}  [Probability that initially there is ONE customer]
+  p2 = 0   {min: 0; max: 1; category: Initial state; caption: two}     [Probability that initially there are TWO customers]
+  p3 = 0   {min: 0; max: 1; category: Initial state; caption: three}   [Probability that initially there are THREE customers]
+
+#output:
+  t     {caption: Time, min}
+  p0    {caption: P(Empty)}
+  busy  {caption: P(Full load)}
+  queue {caption: E(Queue)}
+
+#parameters:
+  arrival = 10   {min: 1; max: 100; category: Means; caption: arrival;  units: min} [Mean arrival time]
+  service = 100  {min: 1; max: 100; category: Means; caption: service; units: min} [Mean service time]`;
+
+const BASIC_MODEL_INPUTS = {
+  _t0: 0,
+  _t1: 60,
+  _h: 1,
+  p0: 1,
+  p1: 0,
+  p2: 0,
+  p3: 0,
+  arrival: 10,
+  service: 100,
+};
+
+const BASIC_MODEL_OUTPUTS = 4;
+
+const CYCLIC_MODEL = `#name: PK-PD
+#tags: model
+#description: Pharmacokinetic-pharmacodynamic (PK-PD) simulation: two-compartment model
+#equations:
+  d(depot)/dt = -KA * depot
+  d(centr)/dt = KA * depot - CL * C2 - Q * C2 + Q * C3
+  d(peri)/dt  = Q * C2 - Q * C3
+  d(eff)/dt  = Kin - Kout * (1 - C2/(EC50 + C2)) * eff
+
+#expressions:
+  C2 = centr / V2
+  C3 = peri / V3
+
+#loop:
+  _count = 10 {caption: count; category: Dosing; min: 1; max: 20} [Number of doses]
+  depot += dose
+
+#argument: t
[…]
