### 2025-11-06 9e67b0f
GROK-19134: Diff Grok: Workflow upd
Added tests to the github workflow
