### 2025-11-06 1434fc8
GROK-19134: Diff Grok: Workflow upd
Updated test workflow
