### 2025-11-06 11400d6
GROK-19134: Diff Grok: MAD upd
Updated MAD
