### 2025-11-06 beb8802
GROK-19134: Diff Grok: Fixed MAD
Fixed the MAD metric
