### 2025-11-20 ecc7d62
GROK-19142: Diff Grok: Docs upd
Updated docs, removed internal features
