### 2025-11-21 bbce162
GROK-19142: Diff Grok Docs: Types description
Added description of types
