+++ index.ts
@@ -1,12 +1,12 @@
 export {Func, ODEs, mrt, ros3prw, ros34prw, CallbackAction, DEFAULT_OPTIONS, SolverOptions,
-  getCallback, SolverMethod} from './src/solver-tools';
+  getCallback, SolverMethod, Callback, IterCheckerCallback, TimeCheckerCallback} from './src/solver-tools';
 
 export {perfProbs, corrProbs, CorrProblem} from './src/examples';
 
 export {DF_NAME, CONTROL_EXPR, MAX_LINE_CHART, getIVP, getScriptLines, getScriptParams, getJScode,
   IVP, DifEqs, Input, SCRIPTING, BRACE_OPEN, BRACE_CLOSE, BRACKET_OPEN, BRACKET_CLOSE, ANNOT_SEPAR,
   CONTROL_SEP, STAGE_COL_NAME, ARG_INPUT_KEYS, DEFAULT_SOLVER_SETTINGS, ModelError,
-  getFunc4worker, Loop, Update, Output} from './src/scripting-tools';
+  getFunc4worker, Loop, Update, Output, Arg} from './src/scripting-tools';
 
 export {IVP2WebWorker, getIvp2WebWorker, solveIvp} from './src/worker-tools';
 
