+++ src/scripting-tools/index.ts
@@ -3,6 +3,6 @@ export {DF_NAME, CONTROL_EXPR, MAX_LINE_CHART} from './constants';
 export {DifEqs, getIVP, getScriptLines, getScriptParams, getJScode, IVP, Input, SCRIPTING,
   BRACE_OPEN, BRACE_CLOSE, BRACKET_OPEN, BRACKET_CLOSE, ANNOT_SEPAR, CONTROL_SEP,
   STAGE_COL_NAME, ARG_INPUT_KEYS, DEFAULT_SOLVER_SETTINGS, getFunc4worker, Loop, Update,
-  Output} from './scripting-tools';
+  Output, Arg} from './scripting-tools';
 
 export {ModelError} from './model-error';
