+++ src/pipeline/output.ts
@@ -180,7 +180,9 @@ function checkApplicability(ivp: IVP): void {
   }
 } // checkApplicability
 
-/** Return a code for output extraction */
+/** Return a code for output extraction
+ * @param ivp - initial value problem
+ */
 export function getOutputNames(ivp: IVP): string[] {
   const outputs = ivp.outputs;
   if (outputs === null)
