### 2025-12-01 ee849d6
GROK-19307: Diff Grok: Contributing docs
Added the contributing stuff
