+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,22 @@
+## Description
+
+Please describe what this PR does and why it is needed.
+
+## Related Issue(s) / Motivation
+
+Closes / Addresses / References: #[issue number] (if applicable)
+
+## Type of change
+
+- [ ] Bug fix  
+- [ ] New feature  
+- [ ] Update / improvement  
+- [ ] Documentation / example update  
+
+## Checklist
+
+- [ ] I have read CONTRIBUTING.md  
+- [ ] My code follows project style (lint + formatting)  
+- [ ] I added / updated tests or examples (if applicable)  
+- [ ] All existing tests pass  
+- [ ] I updated documentation / README (if applicable)  
