+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,35 @@
+
+---
+
+### **.github/ISSUE_TEMPLATE/bug_report.md**
+
+```markdown
+---
+name: Bug report
+about: Report a bug in diff-grok
+title: "[BUG] "
+labels: bug
+assignees: ""
+
+---
+
+**What happened?**  
+A clear and concise description of the problem.
+
+**How to reproduce?**  
+Provide a minimal reproducible example (code snippet, inputs, parameters, Node/TS version, OS).
+
+**What did you expect to happen?**  
+Describe the expected behavior.
+
+**What actually happened?**  
+Describe the actual behavior, error messages, stack trace, wrong output, etc.
+
+**Environment:**  
+- diff-grok version:  
+- Node / TypeScript version:  
+- OS / Platform:  
+- Additional context (e.g. dependency versions):  
+
+(optional) **Example / Test case:**  
+Provide a minimal code snippet or scenario where bug occurs.
