+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,20 @@
+---
+name: Feature request
+about: Suggest a new feature or improvement
+title: "[FEATURE] "
+labels: enhancement
+assignees: ""
+
+---
+
+**What problem would this feature solve?**  
+Describe the motivation, use-cases, or scenarios.
+
+**Proposed solution / API idea (if any):**  
+Give a rough idea how this could look (function signature, parameters, return type, etc.).  
+
+**Alternatives / Workarounds:**  
+Describe any existing workaround or alternative that motivated this.  
+
+**Additional context (optional):**  
+Remarks, performance considerations, backward-compatibility notes, etc.
