### 2025-12-01 75a0e22
GROK-19307: Diff Grok: Bug report upd
Updated the bug report template
