### 2025-12-03 ac441e6
GROK-19328: Diff Grok: Generator
Implemented basic version of synthetic data generator
