+++ src/data-generation/index.ts
@@ -1,0 +1,1 @@
+export {SyntheticDataGenerator} from './syntetic-data-generator';
