+++ src/data-generation/syntetic-data-generator.ts
@@ -1,0 +1,24 @@
+import {getIVP, IVP} from '../scripting-tools';
+import {getIvp2WebWorker, IVP2WebWorker} from '../worker-tools';
+import {InputOpts, ModelInputs, getCaptionToInputNameMap, getDefaultInputs, getInputOptionsMap} from './utils';
+
+export class SyntheticDataGenerator {
+  private ivp: IVP;
+  private ivp2ww: IVP2WebWorker;
+  private inputOpts: Map<string, InputOpts>;
+  private defaultInputs: ModelInputs;
+  private captionToNameMap: Map<string, string>;
+
+  constructor(model: string) {
+    this.ivp = getIVP(model);
+    this.ivp2ww = getIvp2WebWorker(this.ivp);
+    this.inputOpts = getInputOptionsMap(this.ivp);
+    this.defaultInputs = getDefaultInputs(this.inputOpts);
+    this.captionToNameMap = getCaptionToInputNameMap(this.inputOpts);
+
+    console.log(this.ivp);
+    console.log(this.inputOpts);
+    console.log(this.defaultInputs);
+    console.log(this.captionToNameMap);
+  } // constructor
+} // SyntheticDataGenerator
