### 2025-12-03 dcedf58
GROK-19328: Diff Grok: Generator
Implemented the inputs routine
