### 2025-12-04 002fbd0
GROK-19307: Diff Grok: Docs update
Added environments and examples use
