### 2025-12-12 08cc676
GROK-19380: Diff Grok: Performance
Added the reference points
