+++ src/examples/index.ts
@@ -1,9 +1,9 @@
-import {robertson} from './robertson';
-import {hires} from './hires';
-import {vdpol} from './vdpol';
-import {orego} from './orego';
-import {e5} from './e5';
-import {pollution} from './pollution';
+import {robertson, robertsonReferencePoint} from './robertson';
+import {hires, hiresReferencePoint} from './hires';
+import {vdpol, vdpolReferencePoint} from './vdpol';
+import {orego, oregoReferencePoint} from './orego';
+import {e5, e5ReferencePoint} from './e5';
+import {pollution, pollutionReferencePoint} from './pollution';
 
 /** Problems for performance check
  * @internal
@@ -15,6 +15,18 @@ export const perfProbs = [
   orego,
   e5,
   pollution,
+];
+
+/** Reference points for the performance problems
+ * @internal
+ */
+export const refPoints = [
+  robertsonReferencePoint,
+  hiresReferencePoint,
+  vdpolReferencePoint,
+  oregoReferencePoint,
+  e5ReferencePoint,
+  pollutionReferencePoint,
 ];
 
 export {corrProbs, CorrProblem} from './corr-probs';
