+++ src/examples/robertson.ts
@@ -1,7 +1,7 @@
 /** Robertson chemical reaction, updated version (see https://archimede.uniba.it/~testset/problems/rober.php) */
 export const robertson = {
   name: 'Robertson',
-  arg: {name: 't', start: 0, finish: 10e11, step: 2.5e7},
+  arg: {name: 't', start: 0, finish: 10e11, step: 2.5e6},
   initial: [1, 0, 0],
   func: (t: number, y: Float64Array, output: Float64Array) => {
     output[0] = -0.04 * y[0] + 1e4 * y[1] * y[2];
