+++ src/examples/check-methods.ts
@@ -27,23 +27,17 @@ function getError(method: (odes: ODEs) => Float64Array[], corProb: CorrProblem):
   return error;
 } // getError
 
-function getDeviationFromReferencePoint(solution: Float64Array[], refPoint: Float64Array) {
+function getDeviationFromReferencePoint(solution: Float64Array[], refPoint: Float64Array): number {
   let absolute = 0;
-  let relative = 0;
   const lastIdx = solution[0].length - 1;
   let cur = 0;
 
   for (let k = 0; k < solution.length; ++k) {
-    //console.log(solution[k][lastIdx], refPoint[k]);
     cur = Math.abs(solution[k][lastIdx] - refPoint[k]);
     absolute = Math.max(cur, absolute);
-    relative = Math.max(cur / (1 + Math.abs(refPoint[k])), relative);
   }
 
-  return {
-    absolute: absolute,
-    relative: relative,
-  };
+  return absolute;
 } // getDeviationFromReferencePoint
 
 const methods = new Map([
@@ -57,16 +51,14 @@ console.log('Performance\n');
 methods.forEach((method, name) => {
   console.log(' ', name);
 
-  console.log('           PROBLEM   TIME,MS   ABS.ERR*    REL.ERR*');
+  console.log('           PROBLEM   TIME,MS   ERR*');
 
   perfProbs.forEach((odes, idx) => {
     const start = Date.now();
     const solution = method(odes);
-    const errs = getDeviationFromReferencePoint(solution.slice(1), refPoints[idx]);
-    const absErr = errs.absolute.toExponential(2);
-    const relErr = errs.relative.toExponential(2);
+    const absErr = getDeviationFromReferencePoint(solution.slice(1), refPoints[idx]).toExponential(2);
     const finish = Date.now();
-    console.log(`  ${odes.name}     ${finish - start}     ${absErr}     ${relErr}`.padStart(50));
+    console.log(`  ${odes.name}     ${finish - start}     ${absErr}`.padStart(37));
   });
 
   console.log();
@@ -78,7 +70,7 @@ console.log('Correctness (maximum absolute deviation): approximate vs. exact\n')
 
 methods.forEach((method, name) => {
   console.log('  ', name);
-  console.log('          PROBLEM   ABS. ERR');
+  console.log('          PROBLEM   ERR');
 
   corrProbs.forEach((problem) => {
     const error = getError(method, problem);
