+++ src/examples/print-benchmark.ts
@@ -1,0 +1,559 @@
+import {printSolution} from './utils';
+import {robertson} from './robertson';
+import {hires} from './hires';
+import {vdpol} from './vdpol';
+import {orego} from './orego';
+import {e5} from './e5';
+import {pollution} from './pollution';
+import {mrt, ros34prw} from '../solver-tools';
+
+const SEPARATOR = '  '; // use ',' to get csv
+
+/** Print solution of the ROBER benchmark problem
+ * @example
+ * ```ts
+ * // TypeScript usage example
+ * printRobertson();
+ * ```
+ *
+ * ```python
+ * # Python reference calculation for comparison
+ * import numpy as np
+import pandas as pd
+from scipy.integrate import solve_ivp
+
+METHOD = "Radau" # "BDF" or "LSODA"
+
+# -------------------------
+# 1. Define Robertson ODE
+# -------------------------
+def robertson(t, y):
+    A, B, C = y
+    dA = -0.04 * A + 1e4 * B * C
+    dB = 0.04 * A - 1e4 * B * C - 3e7 * B**2
+    dC = 3e7 * B**2
+    return [dA, dB, dC]
+
+# -------------------------
+# 2. Initial conditions
+# -------------------------
+y0 = [1.0, 0.0, 0.0]
+t_start = 0.0
+t_end = 1e11
+t_step = 2.5e6
+t_span = (t_start, t_end)
+
+# Output points
+t_eval = np.arange(t_start, t_end + t_step, t_step)
+
+# -------------------------
+# 3. Solve ODE
+# -------------------------
+sol = solve_ivp(
+    robertson,
+    t_span,
+    y0,
+    method=METHOD,
+    t_eval=t_eval,
+    rtol=1e-7,
+    atol=1e-7
+)
+
+print("Success:", sol.success)
+print("Message:", sol.message)
+
+# sol.t → time array, sol.y.T → concentrations (n_points, 3)
+data = np.column_stack((sol.t, sol.y.T))
+
+# Create DataFrame with headers
+df = pd.DataFrame(data, columns=['time', 'A', 'B', 'C'])
+
+# Save to CSV
+df.to_csv(f'ROBER-using-{METHOD}.csv', index=False)
+```
+ */
+export function printRobertson(): void {
+  printSolution(robertson, mrt, SEPARATOR);
+}
+
+/** Print solution of the HIRES benchmark problem
+ * @example
+ * ```ts
+ * // TypeScript usage example
+ * printHires();
+ * ```
+ *
+ * ```python
+ * # Python reference calculation for comparison
+ * import numpy as np
+import pandas as pd
+from scipy.integrate import solve_ivp
+
+# -------------------------
+# 0. Solver method
+# -------------------------
+METHOD = "Radau"  # stiff solver suitable for HIRES
+
+# -------------------------
+# 1. Define HIRES ODE
+# -------------------------
+def hires(t, y):
+    y1, y2, y3, y4, y5, y6, y7, y8 = y
+    dy = np.zeros(8)
+    dy[0] = -1.71 * y1 + 0.43 * y2 + 8.32 * y3 + 0.0007
+    dy[1] = 1.71 * y1 - 8.75 * y2
+    dy[2] = -10.03 * y3 + 0.43 * y4 + 0.035 * y5
+    dy[3] = 8.32 * y2 + 1.71 * y3 - 1.12 * y4
+    dy[4] = -1.745 * y5 + 0.43 * y6 + 0.43 * y7
[…]
