### 2025-12-16 993a8e4
GROK-19380: Diff Grok: Updated readme
Updated the performance section in readme
