### 2025-12-18 b18fbc5
Readme upd
Updated readme, added gallery
