### 2025-12-18 e52ba4b
Readme: Fixed link
Updated link to the POLL image
