### 2025-12-18 dd3bd0f
Updated readme
Added doi to readme
