### 2025-12-18 99fb598
Updated JOSS paper
Modified the paper with respect to the reviews
