+++ paper-joss/paper.md
@@ -2,7 +2,7 @@
 title: 'Diff Studio: Ecosystem for Interactive Modeling by Ordinary Differential Equations'
 tags:
   - Differential equations
-  - JavaScript
+  - TypeScript
 authors:
   - name: Viktor Makarichev   
     orcid: 0000-0003-1481-9132
@@ -39,153 +39,86 @@ bibliography: paper.bib
 
 # Summary
 
-Ordinary differential equations (ODEs) are crucial in modeling complex systems and phenomena. Their applications range from pharmacology and drug manufacturing 
-to financial modeling and environmental studies.
+Ordinary differential equations (ODEs) are crucial in modeling complex systems and phenomena. Their applications range from pharmacology and drug manufacturing to financial modeling and environmental studies.
 
 **Diff Studio** is a high-performance TypeScript application for solving initial value problems (IVPs) for ODEs directly within web browsers. It consists of two components. The **Diff Grok library** implements numerical methods and formula parsing tools. The **Diff Studio application** integrates Diff Grok tools with **Datagrok**, a scientific computing platform free for personal and academic use.
 
 Diff Studio provides an ecosystem for rapid development of ODE-based applications with reproducible and accessible models.
 
-
 # Statement of need
 
-Scientific modeling of complex processes and phenomena uses ODEs. They are widely applied in diverse fields, including physical processes [@chicone2006ordinary], 
+Scientific modeling of complex processes and phenomena uses ODEs. They are widely applied in diverse fields, including physical processes [@chicone2006ordinary],
 biochemical kinetics [@ingalls2013mathematical], drug delivery systems [@mircioiu2019mathematical], cloud computing [@jafarnejad2019applying], and population dynamics [@hastings2013population].
 
-Analytic methods providing exact solutions can be applied only to a limited class of ODEs. 
+Analytic methods providing exact solutions can be applied only to a limited class of ODEs.
 The use of analytic solutions often proves impractical due to their complexity [@hairer2008solving1]. Numerical methods computing approximate solutions are often preferred.
 
-Many methods for solving ODEs have been recently developed 
-[@hairer2008solving1; @hairer2002solving2]. 
-These methods have been implemented in various software tools, 
+Many methods for solving ODEs have been recently developed
+[@hairer2008solving1; @hairer2002solving2].
+These methods have been implemented in various software tools,
 including libraries and packages for programming languages and
-scientific computing environments. 
-Notable examples include 
-SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials], 
-Julia Differential Equations package [@rackauckas2017differentialequations], 
-SciPy [@2020SciPyNMeth], 
-Maple [@maple2025], 
+scientific computing environments.
+Notable examples include
+SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials],
+Julia Differential Equations package [@rackauckas2017differentialequations],
+SciPy [@2020SciPyNMeth],
+Maple [@maple2025],
 Mathematica [@Mathematica2024],
-Matlab [@MATLAB], 
+Matlab [@MATLAB],
 and deSolve [@soetaert2010solving].
 
-Most tools require expertise, shifting focus from research to development. 
-The goal of this project is to develop an ecosystem providing a combination of a "no-code" approach with comprehensive capabilities for in-browser modeling and analysis.
[…]
