+++ paper-joss/paper.bib
@@ -276,3 +276,27 @@ doi = {10.1137/S1064827594276424},
   year = {2025},
   url = {https://github.com/datagrok-ai/diff-grok?tab=readme-ov-file#performance}
 }
+
+@software{wasm2025,
+  title = {WebAssembly Specification},
+  year = {2025},
+  url = {https://webassembly.github.io/spec/core/}
+}
+
+@software{pyodide2025,
+  title = {Pyodide},
+  year = {2025},
+  url = {https://pyodide.org/}
+}
+
+@software{mathjs,
+  title = {MathJS},
+  year = {2025},
+  url = {https://mathjs.org/}
+}
+
+@software{odexjs,
+  title = {odex-js : ODEX in JavaScript},
+  year = {2025},
+  url = {https://github.com/littleredcomputer/odex-js}
+}
