### 2025-12-18 66462fa
Updated readme
Fixed the badge
