+++ paper-joss/paper.md
@@ -43,95 +43,55 @@ Ordinary differential equations (ODEs) are crucial in modeling complex systems a
 
 **Diff Studio** is a high-performance TypeScript application for solving initial value problems (IVPs) for ODEs directly within web browsers. It consists of two components. The **Diff Grok library** implements numerical methods and formula parsing tools. The **Diff Studio application** integrates Diff Grok tools with **Datagrok**, a scientific computing platform free for personal and academic use.
 
-Diff Studio provides an ecosystem for rapid development of ODE-based applications with reproducible and accessible models.
+Diff Studio is an ecosystem for ODE-based modeling within Datagrok. It enables rapid development of applications with reproducible computations and supports interactive exploration of model behavior. The platform facilitates collaboration through sharing, reuse, and community-driven development.
 
 # Statement of need
 
-Scientific modeling of complex processes and phenomena uses ODEs. They are widely applied in diverse fields, including physical processes [@chicone2006ordinary],
+Scientific modeling of complex processes and phenomena often uses ODEs. They are widely applied in diverse fields, including physical processes [@chicone2006ordinary],
 biochemical kinetics [@ingalls2013mathematical], drug delivery systems [@mircioiu2019mathematical], cloud computing [@jafarnejad2019applying], and population dynamics [@hastings2013population].
 
-Analytic methods providing exact solutions can be applied only to a limited class of ODEs.
-The use of analytic solutions often proves impractical due to their complexity [@hairer2008solving1]. Numerical methods computing approximate solutions are often preferred.
+Analytic methods providing exact solutions can be applied only to a limited class of ODEs. The use of analytic solutions often proves impractical due to their complexity [@hairer2008solving1]. Numerical methods computing approximate solutions are often preferred. Many such methods have been developed [@hairer2008solving1; @hairer2002solving2] and implemented in a variety of software tools, including libraries and packages for programming languages and scientific computing environments. Notable examples include SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials], the Julia Differential Equations package [@rackauckas2017differentialequations], SciPy [@2020SciPyNMeth], Maple [@maple2025], Mathematica [@Mathematica2024], Matlab [@MATLAB], and deSolve [@soetaert2010solving].
 
-Many methods for solving ODEs have been recently developed
-[@hairer2008solving1; @hairer2002solving2].
-These methods have been implemented in various software tools,
-including libraries and packages for programming languages and
-scientific computing environments.
-Notable examples include
-SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials],
-Julia Differential Equations package [@rackauckas2017differentialequations],
-SciPy [@2020SciPyNMeth],
-Maple [@maple2025],
-Mathematica [@Mathematica2024],
-Matlab [@MATLAB],
-and deSolve [@soetaert2010solving].
+For web-based ODE simulation, the following approaches are promising: WebAssembly [@wasm2025], Pyodide [@pyodide2025], and pure JavaScript/TypeScript implementations. WebAssembly provides near-native performance by allowing numerical methods written in C/C++ or Rust to be compiled for the browser, but it requires recompilation whenever equations are updated, limiting its flexibility for iterative model design. Pyodide offers a WebAssembly-based Python distribution that enables the use of NumPy, SciPy, and other scientific libraries, but it incurs large package sizes, performance overhead, and less seamless integration with browser APIs. In contrast, pure JavaScript/TypeScript implementations offer a flexible and performant solution, with libraries such as Math.js [@mathjs] and odex-js [@odexjs], although they require programming proficiency.
 
[…]
