+++ paper-joss/paper.md
@@ -39,45 +39,44 @@ bibliography: paper.bib
 
 # Summary
 
-Ordinary differential equations (ODEs) are crucial in modeling complex systems and phenomena. Their applications range from pharmacology and drug manufacturing to financial modeling and environmental studies.
+Initial value problems (IVPs) for ordinary differential equations (ODEs) are widely used to model complex systems. They are utilized in many domains, including physical processes [@chicone2006ordinary], biochemical kinetics [@ingalls2013mathematical], drug delivery systems [@mircioiu2019mathematical], cloud computing [@jafarnejad2019applying], and population dynamics [@hastings2013population].
 
-**Diff Studio** is a high-performance, interactive TypeScript application for solving initial value problems (IVPs) for ODEs directly within web browsers. It consists of two components. The **Diff Grok library** implements numerical methods and formula parsing tools. The **Diff Studio application** integrates Diff Grok tools with **Datagrok**, a scientific computing platform free for personal and academic use.
+Many well-established numerical solvers for ODEs already exist. However, interactive model exploration, reproducible computation, and collaborative workflows remain difficult to achieve. This is especially true in browser-based environments.
 
-Diff Studio is an ecosystem for ODE-based modeling within Datagrok. It enables rapid development of applications with reproducible computations and supports interactive exploration of model behavior. The platform facilitates collaboration through sharing, reuse, and community-driven development.
+**Diff Studio** is a browser-native environment for solving and exploring IVPs, enabling **interactive**, **collaborative**, and **reproducible** exploration of ODE-based simulations without custom software development or local environment setup.
+
+Diff Studio consists of two components. **Diff Grok** is a high-performance TypeScript library for solving ODE systems defined in a declarative form. **Diff Studio** is a web application that integrates Diff Grok into the **Datagrok** scientific computing platform. Together, they provide autogenerated interfaces and automatic recomputation for browser-based ODE modeling.
 
 # Statement of need
 
-Scientific modeling of complex processes and phenomena often uses ODEs. They are widely applied in diverse fields, including physical processes [@chicone2006ordinary],
-biochemical kinetics [@ingalls2013mathematical], drug delivery systems [@mircioiu2019mathematical], cloud computing [@jafarnejad2019applying], and population dynamics [@hastings2013population].
+ODEs can be solved both analytically and numerically. Analytic methods providing exact solutions can be applied only to a limited class of ODEs. The use of analytic solutions often proves impractical due to their complexity [@hairer2008solving1]. Numerical methods computing approximate solutions are often preferred. Many such methods have been developed [@hairer2008solving1; @hairer2002solving2] and are available in many scientific computing tools. Notable examples include SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials], the Julia Differential Equations package [@rackauckas2017differentialequations], SciPy [@2020SciPyNMeth], Maple [@maple2025], Mathematica [@Mathematica2024], Matlab [@MATLAB], and deSolve [@soetaert2010solving]. These tools provide robust and efficient computations. However, they are mainly designed for desktop or scripting-based workflows.
 
[…]
