### 2026-01-15 eb26154
Paper minor upd
Updated the Summary section of the paper
