### 2026-01-29 dae99b8
Diff Grok paper: Updated text
Updated text with respect to the review
