+++ paper-joss/paper.md
@@ -53,20 +53,20 @@ ODEs can be solved both analytically and numerically. Analytic methods providing
 
 In practice, scientific modeling often requires interactive exploration. Researchers frequently adjust parameters and inspect model behavior iteratively. Collaboration between multiple users is also common. Existing solutions usually require local installation and environment configuration. This makes sharing models and reproducing results difficult. Creating interactive analysis tools often requires substantial programming effort. As a result, attention shifts away from scientific analysis toward software development.
 
-Web-based modeling with ODEs introduces additional challenges. Methods utilizing WebAssembly [@wasm2025] or Pyodide [@pyodide2025] allow reuse of existing numerical libraries. WebAssembly provides near-native performance by allowing numerical methods written in C/C++ or Rust to be compiled for the browser. However, it often requires recompilation when equations or model structure change. This limits its flexibility for iterative model design. Pyodide offers a WebAssembly-based Python distribution that enables the use of NumPy, SciPy, and other scientific libraries. However, it has large package sizes, may incur performance overhead and large bundle sizes, and integrates less seamlessly with browser APIs. In contrast, pure JavaScript/TypeScript implementations offer a flexible and performant solution, with libraries such as Math.js [@mathjs] and odex-js [@odexjs]. They expose low-level APIs, so end users need programming expertise.
-
-Diff Studio addresses these limitations. It offers a low-code interface with extensive functionalities for in-browser modeling and analysis of ODE-based systems.
+Diff Studio addresses these limitations by providing a browser-native environment that combines performance with ease of use. It offers a low-code interface with extensive functionalities for in-browser modeling and analysis of ODE-based systems.
 
 # The solution: Diff Studio
 
-The Diff Grok library provides numerical methods for solving problems specified in a declarative form. It includes:
+Web-based modeling with ODEs introduces additional challenges. Methods utilizing WebAssembly [@wasm2025] or Pyodide [@pyodide2025] allow reuse of existing numerical libraries. WebAssembly provides near-native performance by allowing numerical methods written in C/C++ or Rust to be compiled for the browser. However, it often requires recompilation when equations or model structure change. This limits its flexibility for iterative model design. Pyodide offers a WebAssembly-based Python distribution that enables the use of NumPy, SciPy, and other scientific libraries. However, it has large package sizes, may incur performance overhead, and integrates less seamlessly with browser APIs. In contrast, pure JavaScript/TypeScript implementations offer a flexible and performant solution, with libraries such as Math.js [@mathjs] and odex-js [@odexjs]. They expose low-level APIs, so end users need programming expertise.
 
-- **Solving tools:** numerical methods and computational pipelines. The library implements the modified Rosenbrock triple (MRT) [@Shampine1997], the ROS3PRw [@jax2021], and the ROS34PRw [@rang2015improved] methods. They support both stiff and non-stiff systems. Performance was benchmarked on **Robertson** [@robertson1966solution], **HIRES** [@schafer1975new], **VDPOL** [@vanderpol1926relaxation], **OREGO** [@hairer2002solving2], **E5** [@hairer2002solving2], and **Pollution** [@verwer1994gauss]. Diff Grok allows users to obtain modeling results in near-real time (see \autoref{fig:peformance}). Computational pipelines support multi-stage modeling and solving of IVPs in web workers. This enables parallel computations and analyses, including parameter optimization and sensitivity studies.
[…]
