### 2026-02-02 cb4b6bc
Update paper.md
