+++ paper-joss/paper.md
@@ -45,7 +45,9 @@ Many well-established numerical solvers for ODEs already exist. However, interac
 
 **Diff Studio** is a browser-native environment for solving and exploring IVPs, enabling **interactive**, **collaborative**, and **reproducible** exploration of ODE-based simulations without custom software development or local environment setup.
 
-Diff Studio consists of two components. **Diff Grok** is a high-performance TypeScript library for solving ODE systems defined in a declarative form. **Diff Studio** is a web application that integrates Diff Grok into the **Datagrok** scientific computing platform. Together, they provide autogenerated interfaces and automatic recomputation for browser-based ODE modeling.
+Diff Studio consists of two components:
+1. **Diff Grok**:  general-purpose high-performance TypeScript library for solving ODE systems defined in a declarative form
+2. **Diff Studio**: web application that integrates Diff Grok into the **Datagrok** scientific computing platform. It provides autogenerated user interface, rich interactive visualizations, and the ability to manage ODE models and simulations.
 
 # Statement of need
 
