### 2026-02-02 549be19
Diff Grok paper: Updated text
Polished the text
