+++ paper-joss/paper.md
@@ -41,9 +41,9 @@ bibliography: paper.bib
 
 Ordinary differential equations (ODEs), typically formulated as initial value problems (IVPs), are widely used to model the dynamics of complex systems. They are utilized in many domains, including physical processes [@chicone2006ordinary], biochemical kinetics [@ingalls2013mathematical], drug delivery systems [@mircioiu2019mathematical], cloud computing [@jafarnejad2019applying], and population dynamics [@hastings2013population].
 
-Many well-established numerical solvers for ODEs already exist. However, interactive model exploration, reproducible computation, and collaborative workflows remain difficult to achieve. This is especially true in browser-based environments.
+While a broad ecosystem of mature numerical solvers exists, effective **interactive exploration**, **reproducible computation**, and **collaborative model development** remain difficult to achieve - particularly in browser-based environments. Most existing tools are designed around desktop or scripting workflows and provide limited support for real-time interaction, model sharing, or collaborative analysis.
 
-**Diff Studio** is a browser-native environment for solving and exploring IVPs, enabling **interactive**, **collaborative**, and **reproducible** exploration of ODE-based simulations without custom software development or local environment setup.
+Diff Studio is a **browser-native environment for solving and exploring IVPs**, designed to support interactive, collaborative, and reproducible ODE modeling without custom software development or local environment setup. It enables users to define models declaratively, explore their behavior interactively, and share fully reproducible simulations through the browser.
 
 Diff Studio consists of two components:
 1. **Diff Grok**:  general-purpose high-performance TypeScript library for solving ODE systems defined in a declarative form
@@ -51,47 +51,50 @@ Diff Studio consists of two components:
 
 # Statement of need
 
-ODEs can be solved both analytically and numerically. Analytic methods providing exact solutions can be applied only to a limited class of ODEs. The use of analytic solutions often proves impractical due to their complexity [@hairer2008solving1]. Numerical methods computing approximate solutions are often preferred. Many such methods have been developed [@hairer2008solving1; @hairer2002solving2] and are available in many scientific computing tools. Notable examples include SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials], the Julia Differential Equations package [@rackauckas2017differentialequations], SciPy [@2020SciPyNMeth], Maple [@maple2025], Mathematica [@Mathematica2024], Matlab [@MATLAB], and deSolve [@soetaert2010solving]. These tools provide robust and efficient computations. However, they are mainly designed for desktop or scripting-based workflows.
+ODEs can be solved either analytically or numerically. Analytic methods provide exact solutions but apply only to limited classes of problems and are often impractical due to their complexity [@hairer2008solving1]. Numerical methods, which compute approximate solutions, are therefore the dominant approach. A wide range of such methods has been developed [@hairer2008solving1; @hairer2002solving2] and is available in established scientific computing tools, including SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials], Julia DifferentialEquations [@rackauckas2017differentialequations], SciPy [@2020SciPyNMeth], Maple [@maple2025], Mathematica [@Mathematica2024], Matlab [@MATLAB], and deSolve [@soetaert2010solving].
 
-In practice, scientific modeling often requires interactive exploration. Researchers frequently adjust parameters and inspect model behavior iteratively. Collaboration between multiple users is also common. Existing solutions usually require local installation and environment configuration. This makes sharing models and reproducing results difficult. Creating interactive analysis tools often requires substantial programming effort. As a result, attention shifts away from scientific analysis toward software development.
[…]
