### 2026-02-02 #9090 @vmakarichev
[comparison.pdf](https://github.com/user-attachments/files/25004678/comparison.pdf)
