### 2026-02-13 485b637
GROK-19679: Diff Grok: RK4 method
Implemented the Runge-Kutta 4-th order method
