+++ src/solver-tools/rk4-method.ts
@@ -1,0 +1,248 @@
+/* The Runge-Kutta-Fehlberg 4(5) method (RK4).
+
+   An explicit adaptive Runge-Kutta method of order 4 with an embedded
+   5th-order error estimate for step size control.
+
+   Reference:
+   [1] E. Fehlberg, "Low-order classical Runge-Kutta formulas with stepsize control
+       and their application to some heat transfer problems", NASA Technical Report 315, 1969.
+*/
+
+import {ODEs, max, abs, SAFETY, PSHRNK, PSGROW, REDUCE_COEF, GROW_COEF,
+  ERR_CONTR, TINY, ERROR_MSG} from './solver-defs';
+import {Callback} from './callbacks/callback-base';
+
+// Fehlberg 4(5) Butcher tableau coefficients
+
+// Nodes (c_i)
+const C2 = 1.0 / 4.0;
+const C3 = 3.0 / 8.0;
+const C4 = 12.0 / 13.0;
+// C5 = 1
+const C6 = 1.0 / 2.0;
+
+// Matrix coefficients (a_ij)
+const A21 = 1.0 / 4.0;
+
+const A31 = 3.0 / 32.0;
+const A32 = 9.0 / 32.0;
+
+const A41 = 1932.0 / 2197.0;
+const A42 = -7200.0 / 2197.0;
+const A43 = 7296.0 / 2197.0;
+
+const A51 = 439.0 / 216.0;
+const A52 = -8.0;
+const A53 = 3680.0 / 513.0;
+const A54 = -845.0 / 4104.0;
+
+const A61 = -8.0 / 27.0;
+const A62 = 2.0;
+const A63 = -3544.0 / 2565.0;
+const A64 = 1859.0 / 4104.0;
+const A65 = -11.0 / 40.0;
+
+// 4th-order weights (b_i)
+const B1 = 25.0 / 216.0;
+// B2 = 0
+const B3 = 1408.0 / 2565.0;
+const B4 = 2197.0 / 4104.0;
+const B5 = -1.0 / 5.0;
+// B6 = 0
+
+// Error coefficients: (b_i - bHat_i), where bHat are the 5th-order weights
+const E1 = 1.0 / 360.0;
+// E2 = 0
+const E3 = -128.0 / 4275.0;
+const E4 = -2197.0 / 75240.0;
+const E5 = 1.0 / 50.0;
+const E6 = 2.0 / 55.0;
+
+/** Solve initial value problem using the Runge-Kutta-Fehlberg 4(5) method
+ * @param odes initial value problem for ordinary differential equations
+ * @param callback computations control callback
+ * @returns solution of the problem
+*/
+export function rk4(odes: ODEs, callback?: Callback): Float64Array[] {
+  const f = odes.func;
+
+  // operating variables
+  const t0 = odes.arg.start;
+  const t1 = odes.arg.finish;
+  let h = odes.arg.step;
+  const hDataframe = h;
+  const hMax = hDataframe * 10;
+  const tolerance = odes.tolerance;
+
+  const rowCount = Math.trunc((t1 - t0) / h) + 1;
+  const dim = odes.initial.length;
+
+  // result arrays
+  const tArr = new Float64Array(rowCount);
+  const yArrs = Array<Float64Array>(dim);
+
+  for (let i = 0; i < dim; ++i)
+    yArrs[i] = new Float64Array(rowCount);
+
+  // method routine
+  let timeDataframe = t0 + hDataframe;
+  let t = t0;
+  let tPrev = t0;
+  let hNext = 0.0;
+  let flag = true;
+  let index = 1;
+  let errmax = 0;
+  let hTemp = 0;
+  let tNew = 0;
+
+  // buffers
[…]
