+++ src/solver-tools/index.ts
@@ -2,6 +2,7 @@ export {ODEs, Func} from './solver-defs';
 export {mrt} from './mrt-method';
 export {ros3prw} from './ros3prw-method';
 export {ros34prw} from './ros34prw-method';
+export {rk4} from './rk4-method';
 export {CallbackAction, DEFAULT_OPTIONS, SolverOptions, SolverMethod} from './solver-defs';
 export {getCallback} from './callbacks/callback-tools';
 export {Callback} from './callbacks/callback-base';
