### 2026-02-16 9b2695d
GROK-19679: Diff Grok: Added CC plan
Added plan for implementing methods
