+++ PLAN.md
@@ -1,0 +1,308 @@
+# Plan Template: Adding New ODE Solver Methods
+
+This document provides a step-by-step guide for implementing new numerical solver methods in the diff-grok library.
+
+## Overview
+
+The diff-grok library currently supports:
+- **Implicit methods** (for stiff ODEs): `mrt`, `ros3prw`, `ros34prw`
+- **Explicit methods** (for non-stiff ODEs): `rk4`
+
+All solvers follow a common architecture with adaptive step size control.
+
+## Implementation Checklist
+
+### 1. Create the Solver Implementation File
+
+**Location**: `src/solver-tools/<method-name>-method.ts`
+
+**Structure**:
+```typescript
+/* Method description and references */
+
+import {ODEs, max, abs, SAFETY, PSHRNK, PSGROW, REDUCE_COEF, GROW_COEF,
+  ERR_CONTR, TINY, EPS, tDerivative, jacobian, ERROR_MSG} from './solver-defs';
+import {Callback} from './callbacks/callback-base';
+import {luDecomp, luSolve, solve1d2d} from './lin-alg-tools'; // for implicit methods
+
+// Method-specific constants (Butcher tableau coefficients, gamma values, etc.)
+const COEF_1 = ...;
+const COEF_2 = ...;
+
+/** Solve initial value problem using the <Method Name> method
+ * @param odes initial value problem for ordinary differential equations
+ * @param callback computations control callback
+ * @returns solution of the problem
+*/
+export function <methodName>(odes: ODEs, callback?: Callback): Float64Array[] {
+  // 1. Extract problem parameters
+  const f = odes.func;
+  const t0 = odes.arg.start;
+  const t1 = odes.arg.finish;
+  let h = odes.arg.step;
+  const hDataframe = h;
+  const tolerance = odes.tolerance;
+
+  // For explicit methods: add maximum step size constraint
+  const hMax = hDataframe * 10;  // prevents poor interpolation
+
+  // 2. Pre-allocate result arrays
+  const rowCount = Math.trunc((t1 - t0) / h) + 1;
+  const dim = odes.initial.length;
+  const tArr = new Float64Array(rowCount);
+  const yArrs = Array<Float64Array>(dim);
+  for (let i = 0; i < dim; ++i)
+    yArrs[i] = new Float64Array(rowCount);
+
+  // 3. Initialize working buffers
+  const y = new Float64Array(odes.initial);
+  const yPrev = new Float64Array(odes.initial);
+  const dydt = new Float64Array(dim);
+  const yScale = new Float64Array(dim);
+  const yTemp = new Float64Array(dim);
+  const yErr = new Float64Array(dim);
+
+  // Method-specific buffers (k-vectors, matrices, etc.)
+  // For implicit: W matrix, L/U decomposition, Jacobian
+  // For explicit: k1, k2, k3, ..., kN stage vectors
+
+  // 4. Set initial values
+  tArr[0] = t0;
+  for (let i = 0; i < dim; ++i)
+    yArrs[i][0] = y[i];
+
+  // 5. Main adaptive stepping loop
+  let timeDataframe = t0 + hDataframe;
+  let t = t0;
+  let tPrev = t0;
+  let hNext = 0.0;
+  let flag = true;
+  let index = 1;
+
+  while (flag) {
+    // Compute derivative for scale vector
+    f(t, y, dydt);
+
+    // Callback hook
+    if (callback)
+      callback.onIterationStart();
+
+    // Compute scale vector for error control
+    for (let i = 0; i < dim; ++i)
+      yScale[i] = abs(y[i]) + h * abs(dydt[i]) + TINY;
+
[…]
