### 2026-02-17 b090831
GROK-19679: Diff Grok: Added the AB5 method
Added plan for implementing methods
