+++ src/solver-tools/ab5-method.ts
@@ -1,0 +1,359 @@
+/* The Adams-Bashforth-Moulton predictor-corrector method of order 5 (AB5).
+
+   Uses the 5-step Adams-Bashforth formula as predictor and the 4-step
+   Adams-Moulton formula as corrector in PECE mode (Predict-Evaluate-
+   Correct-Evaluate), with Milne's device for local error estimation.
+   Bootstrapping is performed using the Runge-Kutta-Fehlberg 4(5) method.
+
+   References:
+   [1] E. Hairer, S.P. Norsett, G. Wanner, "Solving Ordinary Differential
+       Equations I: Nonstiff Problems", Springer, 2nd ed., 1993.
+   [2] L.F. Shampine, M.K. Gordon, "Computer Solution of Ordinary Differential
+       Equations: The Initial Value Problem", W.H. Freeman, 1975.
+*/
+
+import {ODEs, max, abs, SAFETY, PSHRNK, PSGROW, REDUCE_COEF, GROW_COEF,
+  ERR_CONTR, TINY, ERROR_MSG} from './solver-defs';
+import {Callback} from './callbacks/callback-base';
+
+// Adams-Bashforth 5-step coefficients (pre-divided by 720)
+// y_{n+1} = y_n + h * (AB0*f_n + AB1*f_{n-1} + AB2*f_{n-2} + AB3*f_{n-3} + AB4*f_{n-4})
+const AB0 = 1901.0 / 720.0;
+const AB1 = -2774.0 / 720.0;
+const AB2 = 2616.0 / 720.0;
+const AB3 = -1274.0 / 720.0;
+const AB4 = 251.0 / 720.0;
+
+// Adams-Moulton 4-step coefficients (pre-divided by 720)
+// y_{n+1} = y_n + h * (AM_NEW*f_{n+1} + AM0*f_n + AM1*f_{n-1} + AM2*f_{n-2} + AM3*f_{n-3})
+const AM_NEW = 251.0 / 720.0;
+const AM0 = 646.0 / 720.0;
+const AM1 = -264.0 / 720.0;
+const AM2 = 106.0 / 720.0;
+const AM3 = -19.0 / 720.0;
+
+// Milne error estimation coefficient for the AB5-AM4 pair: 2/27 = 38/513
+const MILNE = 2.0 / 27.0;
+
+// Number of uniformly-spaced history points required for AB5
+const HIST_LEN = 5;
+
+// Fehlberg 4(5) Butcher tableau coefficients (used during bootstrap phase)
+const C2 = 1.0 / 4.0;
+const C3 = 3.0 / 8.0;
+const C4 = 12.0 / 13.0;
+const C6 = 1.0 / 2.0;
+
+const A21 = 1.0 / 4.0;
+const A31 = 3.0 / 32.0;
+const A32 = 9.0 / 32.0;
+const A41 = 1932.0 / 2197.0;
+const A42 = -7200.0 / 2197.0;
+const A43 = 7296.0 / 2197.0;
+const A51 = 439.0 / 216.0;
+const A52 = -8.0;
+const A53 = 3680.0 / 513.0;
+const A54 = -845.0 / 4104.0;
+const A61 = -8.0 / 27.0;
+const A62 = 2.0;
+const A63 = -3544.0 / 2565.0;
+const A64 = 1859.0 / 4104.0;
+const A65 = -11.0 / 40.0;
+
+const B1 = 25.0 / 216.0;
+const B3 = 1408.0 / 2565.0;
+const B4 = 2197.0 / 4104.0;
+const B5 = -1.0 / 5.0;
+
+const RKE1 = 1.0 / 360.0;
[…]
