+++ src/worker-tools/solving.ts
@@ -1,6 +1,6 @@
 // Tools for solving IVPs in web-workers
 
-import {SolverMethod, mrt, ros3prw, ros34prw, ODEs, getCallback} from '../solver-tools';
+import {SolverMethod, mrt, ros3prw, ros34prw, ODEs, getCallback, rk4, ab5} from '../solver-tools';
 import {Func} from '../solver-tools/solver-defs';
 import {IVP2WebWorker, getFunc} from './scripting';
 
@@ -17,6 +17,12 @@ function getMethod(name: string | undefined): SolverMethod {
 
   case 'ros3prw':
     return ros3prw;
+
+  case 'rk4':
+    return rk4;
+
+  case 'ab5':
+    return ab5;
 
   default:
     return ros34prw;
