### 2026-02-17 ebe870a
GROK-19679: Diff Grok: Updated md files
Updated the readme and plan files
