### 2026-02-17 1df1c1b
GROK-19679: Diff Grok: Implemented Dormand-Prince method
Implemented Dormand-Prince method
