+++ src/solver-tools/rk5-method.ts
@@ -1,0 +1,262 @@
+/* The Dormand-Prince 5(4) method (RK5).
+
+   An explicit adaptive Runge-Kutta method of order 5 with an embedded
+   4th-order error estimate for step size control. Uses the FSAL (First
+   Same As Last) property for efficiency.
+
+   Reference:
+   [1] J.R. Dormand and P.J. Prince, "A family of embedded Runge-Kutta formulae",
+       Journal of Computational and Applied Mathematics, 6(1):19-26, 1980.
+*/
+
+import {ODEs, max, abs, SAFETY, REDUCE_COEF, GROW_COEF,
+  ERR_CONTR, TINY, ERROR_MSG} from './solver-defs';
+import {Callback} from './callbacks/callback-base';
+
+// Step size control exponents for a 5th-order method
+const PSHRNK = -0.2; // -1/p = -1/5
+const PSGROW = -1.0 / 6.0; // -1/(p+1) = -1/6
+
+// Dormand-Prince 5(4) Butcher tableau coefficients
+
+// Nodes (c_i)
+const C2 = 1.0 / 5.0;
+const C3 = 3.0 / 10.0;
+const C4 = 4.0 / 5.0;
+const C5 = 8.0 / 9.0;
+// C6 = 1
+// C7 = 1
+
+// Matrix coefficients (a_ij)
+const A21 = 1.0 / 5.0;
+
+const A31 = 3.0 / 40.0;
+const A32 = 9.0 / 40.0;
+
+const A41 = 44.0 / 45.0;
+const A42 = -56.0 / 15.0;
+const A43 = 32.0 / 9.0;
+
+const A51 = 19372.0 / 6561.0;
+const A52 = -25360.0 / 2187.0;
+const A53 = 64448.0 / 6561.0;
+const A54 = -212.0 / 729.0;
+
+const A61 = 9017.0 / 3168.0;
+const A62 = -355.0 / 33.0;
+const A63 = 46732.0 / 5247.0;
+const A64 = 49.0 / 176.0;
+const A65 = -5103.0 / 18656.0;
+
+// A7i = Bi (FSAL property: 7th-stage coefficients equal the 5th-order weights)
+
+// 5th-order weights (b_i) — used to advance the solution
+const B1 = 35.0 / 384.0;
+// B2 = 0
+const B3 = 500.0 / 1113.0;
+const B4 = 125.0 / 192.0;
+const B5 = -2187.0 / 6784.0;
+const B6 = 11.0 / 84.0;
+// B7 = 0
+
+// Error coefficients: (b_i - bHat_i), where bHat are the 4th-order weights
+const E1 = 71.0 / 57600.0;
+// E2 = 0
+const E3 = -71.0 / 16695.0;
+const E4 = 71.0 / 1920.0;
+const E5 = -17253.0 / 339200.0;
+const E6 = 22.0 / 525.0;
+const E7 = -1.0 / 40.0;
+
+/** Solve initial value problem using the Dormand-Prince 5(4) method
+ * @param odes initial value problem for ordinary differential equations
+ * @param callback computations control callback
+ * @returns solution of the problem
+*/
+export function rk5(odes: ODEs, callback?: Callback): Float64Array[] {
+  const f = odes.func;
+
+  // operating variables
+  const t0 = odes.arg.start;
+  const t1 = odes.arg.finish;
+  let h = odes.arg.step;
+  const hDataframe = h;
+  const hMax = hDataframe * 10;
+  const tolerance = odes.tolerance;
+
+  const rowCount = Math.trunc((t1 - t0) / h) + 1;
+  const dim = odes.initial.length;
+
+  // result arrays
+  const tArr = new Float64Array(rowCount);
[…]
