+++ src/tests/test-defs.ts
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-import {mrt, ros3prw, ros34prw, rk4, ab5} from '../../index';
+import {mrt, ros3prw, ros34prw, rk4, ab5, ab4, rk5} from '../../index';
 
 export const methods = new Map([
   ['MRT', mrt],
@@ -7,6 +7,8 @@ export const methods = new Map([
   ['ROS34PRw', ros34prw],
   ['RK4', rk4],
   ['AB5', ab5],
+  ['AB4', ab4],
+  ['RK5', rk5],
 ]);
 
 /** Implicit methods suitable for stiff performance benchmarks */
