+++ src/solver-tools/ab4-method.ts
@@ -1,0 +1,357 @@
+/* The Adams-Bashforth-Moulton predictor-corrector method of order 4 (AB4).
+
+   Uses the 4-step Adams-Bashforth formula as predictor and the 3-step
+   Adams-Moulton formula as corrector in PECE mode (Predict-Evaluate-
+   Correct-Evaluate), with Milne's device for local error estimation.
+   Bootstrapping is performed using the Runge-Kutta-Fehlberg 4(5) method.
+
+   References:
+   [1] E. Hairer, S.P. Norsett, G. Wanner, "Solving Ordinary Differential
+       Equations I: Nonstiff Problems", Springer, 2nd ed., 1993.
+   [2] L.F. Shampine, M.K. Gordon, "Computer Solution of Ordinary Differential
+       Equations: The Initial Value Problem", W.H. Freeman, 1975.
+*/
+
+import {ODEs, max, abs, SAFETY, PSHRNK, PSGROW, REDUCE_COEF, GROW_COEF,
+  ERR_CONTR, TINY, ERROR_MSG} from './solver-defs';
+import {Callback} from './callbacks/callback-base';
+
+// Adams-Bashforth 4-step coefficients (pre-divided by 24)
+// y_{n+1} = y_n + h * (AB0*f_n + AB1*f_{n-1} + AB2*f_{n-2} + AB3*f_{n-3})
+const AB0 = 55.0 / 24.0;
+const AB1 = -59.0 / 24.0;
+const AB2 = 37.0 / 24.0;
+const AB3 = -9.0 / 24.0;
+
+// Adams-Moulton 3-step coefficients (pre-divided by 24)
+// y_{n+1} = y_n + h * (AM_NEW*f_{n+1} + AM0*f_n + AM1*f_{n-1} + AM2*f_{n-2})
+const AM_NEW = 9.0 / 24.0;
+const AM0 = 19.0 / 24.0;
+const AM1 = -5.0 / 24.0;
+const AM2 = 1.0 / 24.0;
+
+// Milne error estimation coefficient for the AB4-AM3 pair: 19/270
+const MILNE = 19.0 / 270.0;
+
+// Number of uniformly-spaced history points required for AB4
+const HIST_LEN = 4;
+
+// Fehlberg 4(5) Butcher tableau coefficients (used during bootstrap phase)
+const C2 = 1.0 / 4.0;
+const C3 = 3.0 / 8.0;
+const C4 = 12.0 / 13.0;
+const C6 = 1.0 / 2.0;
+
+const A21 = 1.0 / 4.0;
+const A31 = 3.0 / 32.0;
+const A32 = 9.0 / 32.0;
+const A41 = 1932.0 / 2197.0;
+const A42 = -7200.0 / 2197.0;
+const A43 = 7296.0 / 2197.0;
+const A51 = 439.0 / 216.0;
+const A52 = -8.0;
+const A53 = 3680.0 / 513.0;
+const A54 = -845.0 / 4104.0;
+const A61 = -8.0 / 27.0;
+const A62 = 2.0;
+const A63 = -3544.0 / 2565.0;
+const A64 = 1859.0 / 4104.0;
+const A65 = -11.0 / 40.0;
+
+const B1 = 25.0 / 216.0;
+const B3 = 1408.0 / 2565.0;
+const B4 = 2197.0 / 4104.0;
+const B5 = -1.0 / 5.0;
+
+const RKE1 = 1.0 / 360.0;
+const RKE3 = -128.0 / 4275.0;
+const RKE4 = -2197.0 / 75240.0;
+const RKE5 = 1.0 / 50.0;
+const RKE6 = 2.0 / 55.0;
+
+/** Solve initial value problem using the Adams-Bashforth-Moulton predictor-corrector method of order 4
[…]
