+++ src/worker-tools/solving.ts
@@ -1,6 +1,6 @@
 // Tools for solving IVPs in web-workers
 
-import {SolverMethod, mrt, ros3prw, ros34prw, ODEs, getCallback, rk4, ab5} from '../solver-tools';
+import {SolverMethod, mrt, ros3prw, ros34prw, ODEs, getCallback, rk4, ab5, ab4, rkdp} from '../solver-tools';
 import {Func} from '../solver-tools/solver-defs';
 import {IVP2WebWorker, getFunc} from './scripting';
 
@@ -23,6 +23,12 @@ function getMethod(name: string | undefined): SolverMethod {
 
   case 'ab5':
     return ab5;
+
+  case 'ab4':
+    return ab4;
+
+  case 'rkdp':
+    return rkdp;
 
   default:
     return ros34prw;
