+++ README.MD
@@ -17,10 +17,12 @@ A lightweight TypeScript library for solving initial value problem ([IVP](https:
   * The [ROS3PRw](https://doi.org/10.1016/j.cam.2015.03.010) method
   * The [ROS34PRw](https://doi.org/10.1016/j.cam.2015.03.010) method
 * Explicit methods (for non-stiff ODEs):
-  * The Runge-Kutta-Fehlberg 4(5) method ([RK4](https://en.wikipedia.org/wiki/Runge%E2%80%93Kutta_methods))
-  * The Adams-Bashforth predictor-corrector method of order 5 ([AB5](https://en.wikipedia.org/wiki/Linear_multistep_method))
-  * The Adams-Bashforth predictor-corrector method of order 4 ([AB4](https://en.wikipedia.org/wiki/Linear_multistep_method))
-  * The Dormand-Prince 5(4) method ([RK5](https://en.wikipedia.org/wiki/Dormand%E2%80%93Prince_method))
+  * Runge-Kutta type:
+    * The Runge-Kutta-Fehlberg method ([RK4](https://en.wikipedia.org/wiki/Runge%E2%80%93Kutta%E2%80%93Fehlberg_method))
+    * The Dormand-Prince 5(4) method ([RKDP](https://en.wikipedia.org/wiki/Dormand%E2%80%93Prince_method))
+  * Adams-Bashforth-Moulton type:
+    * The predictor-corrector method of order 4 ([AB4](https://en.wikipedia.org/wiki/Linear_multistep_method))
+    * The predictor-corrector method of order 5 ([AB5](https://en.wikipedia.org/wiki/Linear_multistep_method))
 * Scripting:
   * declarative specification of models
   * auto-generated JavaScript code
@@ -77,11 +79,13 @@ on the segment $[t_0, t_1]$ with the step $h$:
    * `ros3prw`- the [ROS3PRw](https://doi.org/10.1016/j.cam.2015.03.010) method
    * `ros34prw` - the [ROS34PRw](https://doi.org/10.1016/j.cam.2015.03.010) method
 
-   Explicit methods (for non-stiff ODEs):
+   Explicit methods (for non-stiff ODEs) — Runge-Kutta type:
    * `rk4` - the Runge-Kutta-Fehlberg 4(5) method
-   * `ab5` - the Adams-Bashforth-Moulton predictor-corrector method of order 5
-   * `ab4` - the Adams-Bashforth-Moulton predictor-corrector method of order 4
-   * `rk5` - the Dormand-Prince 5(4) method
+   * `rkdp` - the Dormand-Prince 5(4) method
+
+   Explicit methods (for non-stiff ODEs) — Adams-Bashforth-Moulton type:
+   * `ab4` - the predictor-corrector method of order 4
+   * `ab5` - the predictor-corrector method of order 5
 
 2. Specify `ODEs` object that defines a problem:
 
