### 2026-02-17 3fb8099
GROK-19679: Diff Grok: Added the RK3 method
Implemented Runge Kutta 3 method
