+++ src/solver-tools/rk3-method.ts
@@ -1,0 +1,219 @@
+/* The Bogacki-Shampine 3(2) method (RK3).
+
+   An explicit adaptive Runge-Kutta method of order 3 with an embedded
+   2nd-order error estimate for step size control. Uses the FSAL (First
+   Same As Last) property for efficiency.
+
+   Reference:
+   [1] P. Bogacki and L.F. Shampine, "A 3(2) pair of Runge-Kutta formulas",
+       Applied Mathematics Letters, 2(4):321-325, 1989.
+*/
+
+import {ODEs, max, abs, SAFETY, REDUCE_COEF, GROW_COEF,
+  ERR_CONTR, TINY, ERROR_MSG} from './solver-defs';
+import {Callback} from './callbacks/callback-base';
+
+// Step size control exponents for a 3rd-order method
+const PSHRNK = -1.0 / 3.0; // -1/p = -1/3
+const PSGROW = -0.25; // -1/(p+1) = -1/4
+
+// Bogacki-Shampine 3(2) Butcher tableau coefficients
+
+// Nodes (c_i)
+const C2 = 1.0 / 2.0;
+const C3 = 3.0 / 4.0;
+// C4 = 1 (FSAL)
+
+// Matrix coefficients (a_ij)
+const A21 = 1.0 / 2.0;
+
+const A32 = 3.0 / 4.0;
+
+// A4i = Bi (FSAL property: 4th-stage coefficients equal the 3rd-order weights)
+
+// 3rd-order weights (b_i) — used to advance the solution
+const B1 = 2.0 / 9.0;
+const B2 = 1.0 / 3.0;
+const B3 = 4.0 / 9.0;
+// B4 = 0
+
+// Error coefficients: (b_i - bHat_i), where bHat are the 2nd-order weights
+const E1 = -5.0 / 72.0;
+const E2 = 1.0 / 12.0;
+const E3 = 1.0 / 9.0;
+const E4 = -1.0 / 8.0;
+
+/** Solve initial value problem using the Bogacki-Shampine 3(2) method
+ * @param odes initial value problem for ordinary differential equations
+ * @param callback computations control callback
+ * @returns solution of the problem
+*/
+export function rk3(odes: ODEs, callback?: Callback): Float64Array[] {
+  const f = odes.func;
+
+  // operating variables
+  const t0 = odes.arg.start;
+  const t1 = odes.arg.finish;
+  let h = odes.arg.step;
+  const hDataframe = h;
+  const hMax = hDataframe * 10;
+  const tolerance = odes.tolerance;
+
+  const rowCount = Math.trunc((t1 - t0) / h) + 1;
+  const dim = odes.initial.length;
+
+  // result arrays
+  const tArr = new Float64Array(rowCount);
+  const yArrs = Array<Float64Array>(dim);
+
+  for (let i = 0; i < dim; ++i)
+    yArrs[i] = new Float64Array(rowCount);
+
+  // method routine
+  let timeDataframe = t0 + hDataframe;
+  let t = t0;
+  let tPrev = t0;
+  let hNext = 0.0;
+  let flag = true;
+  let index = 1;
+  let errmax = 0;
+  let hTemp = 0;
+  let tNew = 0;
+
+  // buffers
+  const y = new Float64Array(odes.initial);
+  const yPrev = new Float64Array(odes.initial);
+  const dydt = new Float64Array(dim);
+  const yScale = new Float64Array(dim);
+  const yTemp = new Float64Array(dim);
+  const yErr = new Float64Array(dim);
+
+  const k1 = new Float64Array(dim);
+  const k2 = new Float64Array(dim);
+  const k3 = new Float64Array(dim);
[…]
