+++ src/solver-tools/index.ts
@@ -6,6 +6,7 @@ export {rk4} from './rk4-method';
 export {ab5} from './ab5-method';
 export {ab4} from './ab4-method';
 export {rkdp} from './rkdp-method';
+export {rk3} from './rk3-method';
 export {CallbackAction, DEFAULT_OPTIONS, SolverOptions, SolverMethod} from './solver-defs';
 export {getCallback} from './callbacks/callback-tools';
 export {Callback} from './callbacks/callback-base';
