+++ src/tests/test-defs.ts
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-import {mrt, ros3prw, ros34prw, rk4, ab5, ab4, rkdp} from '../../index';
+import {mrt, ros3prw, ros34prw, rk4, ab5, ab4, rkdp, rk3} from '../../index';
 
 export const methods = new Map([
   ['MRT', mrt],
@@ -9,6 +9,7 @@ export const methods = new Map([
   ['AB5', ab5],
   ['AB4', ab4],
   ['RKDP', rkdp],
+  ['RK3', rk3],
 ]);
 
 /** Implicit methods suitable for stiff performance benchmarks */
