### 2026-02-18 7036224
GROK-19679: Diff Grok: Lsoda plan
Added lsoda dev plan
