+++ README.MD
@@ -21,7 +21,7 @@ A lightweight TypeScript library for solving initial value problem ([IVP](https:
     * The Bogacki-Shampine 3(2) method ([RK3](https://en.wikipedia.org/wiki/Bogacki%E2%80%93Shampine_method))
     * The Runge-Kutta-Fehlberg method ([RK4](https://en.wikipedia.org/wiki/Runge%E2%80%93Kutta%E2%80%93Fehlberg_method))
     * The Dormand-Prince 5(4) method ([RKDP](https://en.wikipedia.org/wiki/Dormand%E2%80%93Prince_method))
-  * Adams-Bashforth-Moulton type:
+  * Adams-Bashforth type:
     * The predictor-corrector method of order 4 ([AB4](https://en.wikipedia.org/wiki/Linear_multistep_method))
     * The predictor-corrector method of order 5 ([AB5](https://en.wikipedia.org/wiki/Linear_multistep_method))
 * Scripting:
@@ -85,7 +85,7 @@ on the segment $[t_0, t_1]$ with the step $h$:
    * `rk4` - the Runge-Kutta-Fehlberg 4(5) method
    * `rkdp` - the Dormand-Prince 5(4) method
 
-   Explicit methods (for non-stiff ODEs) — Adams-Bashforth-Moulton type:
+   Explicit methods (for non-stiff ODEs) — Adams-Bashforth type:
    * `ab4` - the predictor-corrector method of order 4
    * `ab5` - the predictor-corrector method of order 5
 
