+++ LSODA.md
@@ -1,0 +1,226 @@
+# Plan: Implement LSODA Method
+
+## Context
+
+LSODA (Livermore Solver for Ordinary Differential equations with Automatic method switching) automatically detects stiffness and switches between Adams methods (non-stiff) and BDF methods (stiff). This gives users a single solver that handles both stiff and non-stiff problems without requiring them to know the problem's character in advance.
+
+The diff-grok library already has all the building blocks: Adams predictor-corrector methods (AB4/AB5), implicit solver infrastructure (Jacobian, LU decomposition), and adaptive step control. LSODA unifies these under automatic stiffness detection.
+
+**References:**
+- Hindmarsh, A.C. and Petzold, L.R., "LSODA, Ordinary Differential Equation Solver for Stiff or Non-Stiff System", ODEPACK, 1983.
+- Hairer, E. and Wanner, G., "Solving Ordinary Differential Equations II: Stiff and Differential-Algebraic Problems", Springer, 1996.
+
+## Design Decisions
+
+- **Fixed-order** approach (Adams order 4 + BDF order 2) rather than full variable-order Nordsieck arrays — keeps implementation consistent with existing codebase patterns (~400 lines vs ~1500+)
+- **Non-stiff mode**: Adams-Bashforth-Moulton order 4 (identical to existing AB4 method), bootstrapped with RKF45
+- **Stiff mode**: BDF2 with Modified Newton iteration, using existing `jacobian()`, `luDecomp()`, `luSolve()`, `solve1d2d()`
+- **Stiffness detection**: consecutive step rejection counter (Adams→BDF), periodic trial explicit step (BDF→Adams)
+- **Classification**: hybrid method — added to both `methods` (all correctness tests) and `implicitMethods` (stiff performance tests)
+
+---
+
+## Step 1: Create `src/solver-tools/lsoda-method.ts`
+
+### Constants
+
+**Adams-Bashforth-Moulton order 4** (same as [ab4-method.ts](src/solver-tools/ab4-method.ts)):
+```
+AB: 55/24, -59/24, 37/24, -9/24
+AM: 9/24, 19/24, -5/24, 1/24
+MILNE = 19/270
+ADAMS_HIST_LEN = 4
+```
+
+**BDF2**:
+```
+y_{n+1} = (4/3)*y_n - (1/3)*y_{n-1} + (2/3)*h*f(t_{n+1}, y_{n+1})
+BDF2_A0 = 4/3, BDF2_A1 = -1/3, BDF2_GAMMA = 2/3
+```
+
+**BDF2 adaptive step exponents** (order 2):
+```
+BDF_PSHRNK = -0.5       (= -1/p)
+BDF_PSGROW = -1/3        (= -1/(p+1))
+```
+
+**Newton iteration**:
+```
+MAX_NEWTON_ITER = 4
+NEWTON_TOL = 0.33         (convergence rate threshold)
+```
+
+**Stiffness detection**:
+```
+STIFF_DETECT_THRESHOLD = 5    (consecutive Adams rejects -> switch to BDF)
+NONSTIFF_TEST_INTERVAL = 30   (accepted BDF steps -> trial Adams step)
+```
+
+**RKF45 Butcher tableau** (same as [ab4-method.ts](src/solver-tools/ab4-method.ts) lines 40-70).
+
+### Buffer Allocation
+
+Pre-allocate all `Float64Array` buffers (zero GC pressure):
+
+| Category | Buffers |
+|----------|---------|
+| Common | `y`, `yPrev`, `dydt`, `yScale`, `yTemp`, `yErr` |
[…]
