### 2026-02-18 d1e593f
GROK-19679: Diff Grok: Lsoda method
Implemented LSODA, several approaches
