+++ src/solver-tools/solver-defs.ts
@@ -158,6 +158,7 @@ export enum ERROR_MSG {
   AB4_FAILS = 'The Adams-Bashforth-Moulton 4 method fails',
   RKDP_FAILS = 'The Dormand-Prince 5(4) method fails',
   RK3_FAILS = 'The Bogacki-Shampine 3(2) method fails',
+  LSODA_FAILS = 'The LSODA method fails',
 };
 
 /** Callback action
