+++ src/solver-tools/lsoda-method.ts
@@ -1,0 +1,1571 @@
+/* LSODA — Livermore Solver for Ordinary Differential Equations with
+   Automatic method switching between Adams-Moulton (non-stiff, orders 1–12)
+   and BDF (stiff, orders 1–5), using the Nordsieck array representation.
+
+   References:
+   [1] Hindmarsh, A.C. and Petzold, L.R., "LSODA, Ordinary Differential
+       Equation Solver for Stiff or Non-Stiff System", ODEPACK, 1983.
+   [2] Byrne, G.D. and Hindmarsh, A.C., "A Polyalgorithm for the Numerical
+       Solution of Ordinary Differential Equations", ACM TOMS, 1(1), 1975.
+   [3] Hairer, E. and Wanner, G., "Solving Ordinary Differential Equations
+       II: Stiff and Differential-Algebraic Problems", Springer, 1996.
+   [4] Radhakrishnan, K. and Hindmarsh, A.C., "Description and Use of LSODE,
+       the Livermore Solver for Ordinary Differential Equations", 1993.
+*/
+
+import {ODEs, max, abs, SAFETY, PSHRNK, PSGROW, REDUCE_COEF, GROW_COEF,
+  ERR_CONTR, TINY, EPS, jacobian, ERROR_MSG} from './solver-defs';
+import {Callback} from './callbacks/callback-base';
+import {luDecomp, luSolve, solve1d2d} from './lin-alg-tools';
+
+// ================================================================
+//  MAXIMUM ORDERS
+// ================================================================
+
+const ADAMS_MAX_ORDER = 12;
+const BDF_MAX_ORDER = 5;
+
+// ================================================================
+//  NEWTON ITERATION
+// ================================================================
+
+const MAX_NEWTON_ITER = 7;
+const NEWTON_CONV_TOL = 0.33;
+
+// ================================================================
+//  STIFFNESS DETECTION
+// ================================================================
+
+const STIFF_DETECT_THRESHOLD = 5;
+const NONSTIFF_TEST_INTERVAL = 100;
+
+// ================================================================
+//  STEP SIZE ADAPTATION
+// ================================================================
+
+const MAX_STEP_GROW = 5.0;
+const MIN_STEP_SHRINK = 0.2;
+
+// ================================================================
+//  MODE SWITCH SAFEGUARDS
+// ================================================================
+
+const MAX_SWITCH_COOLDOWN = 3;
+const MAX_SWITCH_GROW = 2.0;
+
+// ================================================================
+//  JACOBIAN REFRESH
+// ================================================================
+
+const MAX_JACOBIAN_AGE = 20;
+
+// ================================================================
+//  REPEATED FAILURE THRESHOLDS
+// ================================================================
+
+const MAX_CONSECUTIVE_FAILS = 30;
+const FORCE_JAC_RECOMPUTE_FAILS = 3;
+const ORDER_DROP_FAILS = 7;
+
+// ================================================================
+//  RKF45 BUTCHER TABLEAU (for bootstrap & trial steps)
+// ================================================================
+
+const RK_C2 = 1.0 / 4.0;
+const RK_C3 = 3.0 / 8.0;
+const RK_C4 = 12.0 / 13.0;
[…]
