+++ src/solver-tools/index.ts
@@ -7,6 +7,7 @@ export {ab5} from './ab5-method';
 export {ab4} from './ab4-method';
 export {rkdp} from './rkdp-method';
 export {rk3} from './rk3-method';
+export {lsodaWeb, lsoda} from './lsoda-method';
 export {CallbackAction, DEFAULT_OPTIONS, SolverOptions, SolverMethod} from './solver-defs';
 export {getCallback} from './callbacks/callback-tools';
 export {Callback} from './callbacks/callback-base';
