### 2026-02-19 6cbcead
GROK-19679: Diff Grok: LSODA tests upd
Removed LSODA from the performance tests
