### 2026-02-19 3084ec7
GROK-19679: Diff Grok: Removed debugging tests
Removed debugging tests for LSODA
