### 2026-02-19 151eed1
GROK-19679: Diff Grok: MRT upd
Updated the MRT method
