### 2026-02-19 7ce28a3
GROK-19679: Diff Grok: MRT step size upd
Updated adaptive step size control in the MRT method
