### 2026-02-19 0bf2cc3
GROK-19679: Diff Grok: MRT init step upd
Updated initialization of the step in the MRT method
