+++ INIT.md
@@ -1,0 +1,171 @@
+# Initial Step Size Selection for the MRT Method
+
+**Context:** The MRT solver uses a single scalar `tolerance` parameter with the error control scheme:
+
+$$\frac{|e_i|}{\mathrm{sc}_i} \leq \mathrm{tolerance}, \quad \mathrm{sc}_i = |y_i| + h|f_i| + \mathrm{TINY}$$
+
+where `TINY` is a machine-small constant preventing division by zero.
+
+This document describes the standard algorithm for selecting the initial step size $h_0$, adapted to this error control convention.
+
+**Reference:** E. Hairer, S.P. Nørsett, G. Wanner, *Solving Ordinary Differential Equations I*, Springer, 1993, Section II.4.
+
+---
+
+## Algorithm
+
+Given:
+
+- Initial point $(t_0, y_0)$, system dimension $m$
+- Right-hand side $f(t, y)$
+- Scalar tolerance `tol` (≈ `rtol`)
+- Constant `TINY` (≈ `atol`)
+- Method order $p = 2$
+
+---
+
+### Step 1. Compute the scale vector
+
+$$\mathrm{sc}_i = |y_{0,i}| + \mathrm{TINY}, \quad i = 1, \ldots, m$$
+
+> **Note:** At the initial point, we do not yet have $h$, so the term $h|f_i|$ from the runtime
+> error control is omitted. This is consistent with Hairer's approach, where the scale
+> is based on the solution magnitude alone.
+
+---
+
+### Step 2. Evaluate the right-hand side
+
+$$f_0 = f(t_0, y_0)$$
+
+---
+
+### Step 3. Compute norms of the initial data
+
+$$d_0 = \sqrt{\frac{1}{m} \sum_{i=1}^{m} \left(\frac{y_{0,i}}{\mathrm{sc}_i}\right)^2}$$
+
+$$d_1 = \sqrt{\frac{1}{m} \sum_{i=1}^{m} \left(\frac{f_{0,i}}{\mathrm{sc}_i}\right)^2}$$
+
+> These are RMS norms of the initial state and the initial derivative, scaled by `sc`.
+> The $L_\infty$ norm (consistent with the `max`-based error control in the code) can also be used:
+>
+> $$d_0 = \max_i \frac{|y_{0,i}|}{\mathrm{sc}_i}, \qquad d_1 = \max_i \frac{|f_{0,i}|}{\mathrm{sc}_i}$$
+
+---
+
+### Step 4. First guess for $h_0$
+
+$$h_0 = \begin{cases} \dfrac{0.01 \cdot d_0}{d_1}, & \text{if } d_0 \geq 10^{-5} \text{ and } d_1 \geq 10^{-5}, \\[8pt] 10^{-6}, & \text{otherwise}. \end{cases}$$
+
+The ratio $d_0 / d_1$ estimates the time scale on which $y$ changes by a fraction of itself. The factor $0.01$ ensures a conservative start.
+
+---
+
+### Step 5. Explicit Euler probe
+
+Perform one explicit Euler step of size $h_0$ to estimate the second derivative:
+
+$$y_1 = y_0 + h_0 \, f_0$$
+
+$$f_1 = f(t_0 + h_0, \; y_1)$$
+
+$$d_2 = \frac{1}{h_0}\sqrt{\frac{1}{m} \sum_{i=1}^{m} \left(\frac{f_{1,i} - f_{0,i}}{\mathrm{sc}_i}\right)^2}$$
+
+> $d_2$ approximates $\|y''\| / \mathrm{sc}$ — the scaled norm of the second derivative.
+
+---
+
+### Step 6. Refine the guess using the method order
+
[…]
