### 2026-02-19 0617def
GROK-19679: Diff Grok: LSODA
Added the LSODA solver
