### 2026-02-20 90464a3
GROK-19679: Diff Grok: Fixed dim bug
Fixed minor bug related to dimension
