### 2026-02-20 a8ba318
GROK-19679: Diff Grok: Added lsoda lib
Added lsoda library
