+++ src/worker-tools/solving.ts
@@ -1,6 +1,6 @@
 // Tools for solving IVPs in web-workers
 
-import {SolverMethod, mrt, ros3prw, ros34prw, ODEs, getCallback, rk4, ab5, ab4, rkdp, rk3, lsoda} from '../solver-tools';
+import {SolverMethod, mrt, ros3prw, ros34prw, ODEs, getCallback, rk4, ab5, ab4, rkdp, rk3} from '../solver-tools';
 import {Func} from '../solver-tools/solver-defs';
 import {IVP2WebWorker, getFunc} from './scripting';
 
@@ -32,9 +32,6 @@ function getMethod(name: string | undefined): SolverMethod {
 
   case 'rk3':
     return rk3;
-
-  case 'lsoda':
-    return lsoda;
 
   default:
     return ros34prw;
