+++ src/solver-tools/lsoda/lsoda.ts
@@ -1,11 +1,24 @@
 import {
-  LsodaContext, LsodaCommon, LsodaOpt, LsodaFunc, ETA,
+  LsodaContext, LsodaCommon, LsodaOpt, LsodaFunc, NordsieckSnapshot, ETA,
 } from './common';
-import { vmnorm } from './blas';
-import { intdy } from './intdy';
-import { stoda } from './stoda';
+import {vmnorm} from './blas';
+import {intdy} from './intdy';
+import {stoda} from './stoda';
+import {DenseOutput} from './dense';
 
 // ---- Internal helpers ----
+
+function captureSnapshot(c: LsodaCommon, neq: number): NordsieckSnapshot {
+  const nq = c.nq;
+  const yh: Float64Array[] = new Array(nq + 1);
+  for (let j = 0; j <= nq; j++) {
+    const row = new Float64Array(neq);
+    for (let i = 0; i < neq; i++)
+      row[i] = c.yh[j + 1][i + 1]; // convert 1-based to 0-based
+    yh[j] = row;
+  }
+  return {tn: c.tn, h: c.h, hu: c.hu, nq, yh};
+}
 
 function ewset(ctx: LsodaContext, ycur: Float64Array): void {
   const c = ctx.common!;
@@ -168,9 +181,9 @@ export function lsodaReset(ctx: LsodaContext): void {
  * Free context resources. In JS/TS, this is mostly a no-op (GC handles it).
  */
 export function lsodaFree(ctx: LsodaContext): void {
-  if (ctx.error) {
+  if (ctx.error)
     console.error(`unhandled error message: ${ctx.error}`);
-  }
+
   ctx.common = null;
 }
 
@@ -195,7 +208,7 @@ export function lsoda(
   function hardfailure(msg: string): { t: number; state: number } {
     ctx.error = msg;
     ctx.state = -3;
-    return { t, state: ctx.state };
+    return {t, state: ctx.state};
   }
 
   function softfailure(code: number, msg: string): { t: number; state: number } {
@@ -204,18 +217,19 @@ export function lsoda(
       y[i] = c.yh[1][i];
     t = c.tn;
     ctx.state = code;
-    return { t, state: ctx.state };
+    return {t, state: ctx.state};
   }
 
   function successreturn(): { t: number; state: number } {
     for (let i = 1; i <= neq; i++)
       y[i] = c.yh[1][i];
     t = c.tn;
-    if (itask === 4 || itask === 5)
+    if (itask === 4 || itask === 5) {
       if (ihit)
         t = tcrit;
+    }
     ctx.state = 2;
-    return { t, state: ctx.state };
+    return {t, state: ctx.state};
   }
 
   function intdyreturn(): { t: number; state: number } {
@@ -228,12 +242,12 @@ export function lsoda(
     }
     t = tout;
     ctx.state = 2;
-    return { t, state: ctx.state };
+    return {t, state: ctx.state};
   }
 
-  if (c === null) {
+  if (c === null)
     return hardfailure('[lsoda] illegal common block did you call lsoda_prepare?');
-  }
+
 
   // Block b: initial/continuation call setup
   let h0: number = 0;
@@ -244,17 +258,16 @@ export function lsoda(
   if (ctx.state === 1 || ctx.state === 3) {
     h0 = opt.h0;
     if (ctx.state === 1) {
-      if ((tout - t) * h0 < 0.) {
+      if ((tout - t) * h0 < 0.)
         return hardfailure(`[lsoda] tout = ${tout} behind t = ${t}. integration direction is given by ${h0}`);
-      }
     }
   }
 
   const itask = opt.itask;
 
-  if (ctx.state === 3) {
+  if (ctx.state === 3)
     jstart = -1;
-  }
+
 
   // Block c: initial call only (state == 1)
[…]
