+++ src/solver-tools/lsoda/index.ts
@@ -1,4 +1,5 @@
-export { Lsoda, lsoda, lsodaPrepare, lsodaReset, lsodaFree } from './lsoda';
-export type { OdeFunction, LsodaSolveResult } from './lsoda';
-export type { LsodaOpt, LsodaFunc } from './common';
-export { LsodaContext } from './common';
+export {Lsoda, lsoda, lsodaPrepare, lsodaReset, lsodaFree} from './lsoda';
+export type {OdeFunction, LsodaSolveResult} from './lsoda';
+export type {LsodaOpt, LsodaFunc, NordsieckSnapshot} from './common';
+export {LsodaContext} from './common';
+export {DenseOutput} from './dense';
