+++ src/tests/performance.test.ts
@@ -4,12 +4,14 @@ import {perfProbs} from '../../index';
 import {implicitMethods, TIMEOUT_MS} from './test-defs';
 
 implicitMethods.forEach((method, name) => {
-  perfProbs.forEach((odes) => {
-    test(`Performance: method - ${name}, problem - ${odes.name}`, () => {
-      const start = performance.now();
-      method(odes);
-      const elapsed = performance.now() - start;
-      expect(elapsed).toBeLessThan(TIMEOUT_MS);
+  describe(`${name}`, () => {
+    perfProbs.forEach((odes) => {
+      test(`${odes.name}`, () => {
+        const start = performance.now();
+        method(odes);
+        const elapsed = performance.now() - start;
+        expect(elapsed).toBeLessThan(TIMEOUT_MS);
+      });
     });
   });
 });

[…]
