### 2026-02-20 4000d7f
GROK-19679: Diff Grok: Updated the LSODA solver
Updated the lsoda method
