+++ src/solver-tools/lsoda-method.ts
@@ -1,0 +1,136 @@
+/* LSODA library wrapper method.
+
+   Wraps the variable-order Nordsieck-based LSODA solver into the standard SolverMethod interface.
+   Automatically switches between Adams (non-stiff) and BDF (stiff) methods.
+
+   References:
+   [1] Hindmarsh, A.C. "ODEPACK, A Systematized Collection of ODE Solvers."
+   [2] Petzold, L.R. "Automatic Selection of Methods for Solving Stiff and Nonstiff
+       Systems of Ordinary Differential Equations."
+*/
+
+import {ODEs, ERROR_MSG} from './solver-defs';
+import {Callback} from './callbacks/callback-base';
+import {Lsoda} from './lsoda';
+
+/** Wraps an ODEs func into the OdeFunction signature compatible with Lsoda (returns 0). */
+function wrapFunc(f: ODEs['func']): (t: number, y: Float64Array, ydot: Float64Array) => number {
+  return (t, y, ydot) => {
+    f(t, y, ydot);
+    return 0;
+  };
+}
+
+/** Solve initial value problem using the LSODA library method
+ * @param odes initial value problem for ordinary differential equations
+ * @param callback computations control callback
+ * @returns solution of the problem
+ * @example
+ * // Example. Solve the following initial value problem using the LSODA method:
+ * //   dx/dt = x + y - t
+ * //   dy/dt = x * y + t
+ * //   x(0) = 1
+ * //   y(0) = -1
+ * // on [0, 2] with the step 0.01.
+ * import { ODEs, lsoda } from 'diff-grok';
+ *
+ * const task: ODEs = {
+ *   name: 'Example', // name of your model
+ *   arg: {
+ *       name: 't',  // name of the argument
+ *       start: 0,   // initial value of the argument
+ *       finish: 2,  // final value of the argument
+ *       step: 0.01, // solution grid step
+ *   },
+ *   initial: [1, -1], // initial values
+ *   func: (t: number, y: Float64Array, output: Float64Array) => { // right-hand side of the system
+ *     output[0] = y[0] + y[1] - t; // 1-st equation
+ *     output[1] = y[0] * y[1] + t; // 2-nd equation
+ *   },
+ *   tolerance: 1e-7, // tolerance
+ *   solutionColNames: ['x', 'y'], // names of solution functions
+ * };
+ *
+ * try {
+ *   // Solve the problem using the LSODA method
+ *   const solution = lsoda(task);
+ *
+ *   // Print a table with the results
+ *   console.log(task.arg.name, task.solutionColNames[0], task.solutionColNames[1]);
+ *   const length = solution[0].length;
+ *   for (let i = 0; i < length; ++i)
+ *     console.log(solution[0][i], solution[1][i], solution[2][i]);
+ * } catch (err) {
+ *   console.log('Solver failed: ', err instanceof Error ? err.message : 'Unknown problem!');
+ * }
+*/
+export function lsoda(odes: ODEs, callback?: Callback): Float64Array[] {
+  // 1. Extract problem parameters
+  const t0 = odes.arg.start;
+  const t1 = odes.arg.finish;
+  const step = odes.arg.step;
+  const tolerance = odes.tolerance;
+  const dim = odes.initial.length;
+  const mxstep = 50000;
+
+  // 2. Create LSODA solver with dense output enabled
+  const rtol = new Float64Array(dim).fill(tolerance);
+  const atol = new Float64Array(dim).fill(tolerance);
+
+  const solver = new Lsoda(wrapFunc(odes.func), dim, {
[…]
