### 2026-02-23 ad919ab
GROK-19679: Diff Grok: Add bioreactor
Added the bioreactor example
