+++ src/worker-tools/solving.ts
@@ -34,11 +34,11 @@ function getMethod(name: string | undefined): SolverMethod {
   case 'rk3':
     return rk3;
 
-  case 'lsoda':
-    return lsoda;
+  case 'ros34prw':
+    return ros34prw;
 
   default:
-    return ros34prw;
+    return lsoda;
   }
 }
 
