### 2026-02-23 e1d1f1a
GROK-19679: Diff Grok: Refactoring
Refactored the code
