+++ PLAN.md
@@ -1,394 +1,0 @@
-# Plan Template: Adding New ODE Solver Methods
-
-This document provides a step-by-step guide for implementing new numerical solver methods in the diff-grok library.
-
-## Overview
-
-The diff-grok library currently supports:
-- **Implicit methods** (for stiff ODEs): `mrt`, `ros3prw`, `ros34prw`
-- **Explicit methods** (for non-stiff ODEs): `rk4`
-
-All solvers follow a common architecture with adaptive step size control.
-
-## Implementation Checklist
-
-### 1. Create the Solver Implementation File
-
-**Location**: `src/solver-tools/<method-name>-method.ts`
-
-**Structure**:
-```typescript
-/* Method description and references */
-
-import {ODEs, max, abs, SAFETY, PSHRNK, PSGROW, REDUCE_COEF, GROW_COEF,
-  ERR_CONTR, TINY, EPS, tDerivative, jacobian, ERROR_MSG} from './solver-defs';
-import {Callback} from './callbacks/callback-base';
-import {luDecomp, luSolve, solve1d2d} from './lin-alg-tools'; // for implicit methods
-
-// Method-specific constants (Butcher tableau coefficients, gamma values, etc.)
-const COEF_1 = ...;
-const COEF_2 = ...;
-
-/** Solve initial value problem using the <Method Name> method
- * @param odes initial value problem for ordinary differential equations
- * @param callback computations control callback
- * @returns solution of the problem
- * @example
- * // Example. Solve the following initial value problem using the <Method Name> method:
- * //   dx/dt = x + y - t
- * //   dy/dt = x * y + t
- * //   x(0) = 1
- * //   y(0) = -1
- * // on [0, 2] with the step 0.01.
- * import { ODEs, <methodName> } from 'diff-grok';
- *
- * const task: ODEs = {
- *   name: 'Example', // name of your model
- *   arg: {
- *       name: 't',  // name of the argument
- *       start: 0,   // initial value of the argument
- *       finish: 2,  // final value of the argument
- *       step: 0.01, // solution grid step
- *   },
- *   initial: [1, -1], // initial values
- *   func: (t: number, y: Float64Array, output: Float64Array) => { // right-hand side of the system
- *     output[0] = y[0] + y[1] - t; // 1-st equation
- *     output[1] = y[0] * y[1] + t; // 2-nd equation
- *   },
- *   tolerance: 1e-7, // tolerance
- *   solutionColNames: ['x', 'y'], // names of solution functions
- * };
- *
- * try {
- *   // Solve the problem using the <Method Name> method
- *   const solution = <methodName>(task);
- *
- *   // Print a table with the results
- *   console.log(task.arg.name, task.solutionColNames[0], task.solutionColNames[1]);
- *   const length = solution[0].length;
- *   for (let i = 0; i < length; ++i)
- *     console.log(solution[0][i], solution[1][i], solution[2][i]);
- * } catch (err) {
- *   console.log('Solver failed: ', err instanceof Error ? err.message : 'Unknown problem!');
- * }
-*/
-export function <methodName>(odes: ODEs, callback?: Callback): Float64Array[] {
-  // 1. Extract problem parameters
-  const f = odes.func;
-  const t0 = odes.arg.start;
-  const t1 = odes.arg.finish;
-  let h = odes.arg.step;
-  const hDataframe = h;
-  const tolerance = odes.tolerance;
-
-  // For explicit methods: add maximum step size constraint
[…]
