+++ docs/media/print-benchmark.ts
@@ -1,0 +1,594 @@
+import {robertson} from './robertson';
+import {hires} from './hires';
+import {vdpol} from './vdpol';
+import {orego} from './orego';
+import {e5} from './e5';
+import {pollution} from './pollution';
+import {ODEs, SolverMethod, mrt, ros34prw} from '../../index';
+
+/** Solve the given problem and output to console the solution */
+export function printSolution(task: ODEs, method: SolverMethod, separator: string): void {
+  try {
+  // Solve the problem
+    const solution = method(task);
+
+    // Output results
+
+    console.log(task.name);
+
+    let line: string = [task.arg.name].concat(task.solutionColNames).join(separator);
+
+    console.log(line);
+
+    const length = solution[0].length;
+
+    for (let i = 0; i < length; ++i) {
+      line = solution.map((col) => col[i].toString()).join(separator);
+      console.log(line);
+    }
+  } catch (err) {
+    console.log('Solver failed: ', err instanceof Error ? err.message : 'Unknown problem!');
+  }
+}
+
+
+const SEPARATOR = '  '; // use ',' to get csv
+
+/** Print solution of the ROBER benchmark problem
+ * @example
+ * ```ts
+ * // TypeScript usage example
+ * import {printRobertson} from 'diff-grok';
+ *
+ * printRobertson();
+ * ```
+ *
+ * ```python
+ * # Python reference calculation for comparison
+ * import numpy as np
+import pandas as pd
+from scipy.integrate import solve_ivp
+
+METHOD = "Radau" # "BDF" or "LSODA"
+
+# -------------------------
+# 1. Define Robertson ODE
+# -------------------------
+def robertson(t, y):
+    A, B, C = y
+    dA = -0.04 * A + 1e4 * B * C
+    dB = 0.04 * A - 1e4 * B * C - 3e7 * B**2
+    dC = 3e7 * B**2
+    return [dA, dB, dC]
+
+# -------------------------
+# 2. Initial conditions
+# -------------------------
+y0 = [1.0, 0.0, 0.0]
+t_start = 0.0
+t_end = 1e11
+t_step = 2.5e6
+t_span = (t_start, t_end)
+
+# Output points
+t_eval = np.arange(t_start, t_end + t_step, t_step)
+
+# -------------------------
+# 3. Solve ODE
+# -------------------------
+sol = solve_ivp(
+    robertson,
+    t_span,
+    y0,
+    method=METHOD,
+    t_eval=t_eval,
+    rtol=1e-7,
+    atol=1e-7
+)
+
+print("Success:", sol.success)
+print("Message:", sol.message)
+
+# sol.t → time array, sol.y.T → concentrations (n_points, 3)
+data = np.column_stack((sol.t, sol.y.T))
+
+# Create DataFrame with headers
+df = pd.DataFrame(data, columns=['time', 'A', 'B', 'C'])
+
+# Save to CSV
+df.to_csv(f'ROBER-using-{METHOD}.csv', index=False)
+```
+ */
+export function printRobertson(): void {
+  printSolution(robertson, mrt, SEPARATOR);
+}
+
+/** Print solution of the HIRES benchmark problem
+ * @example
+ * ```ts
+ * // TypeScript usage example
+ * import {printHires} from 'diff-grok';
+ *
[…]
