### 2026-02-27 7c7f724
GROK-19679: Diff Grok: License features
Updated the license features for ported libs
