+++ src/solver-tools/lsoda/dense.ts
@@ -1,3 +1,15 @@
+/**
+ * TypeScript port of liblsoda (https://github.com/sdwfrost/liblsoda)
+ * Original LSODA algorithm by Linda Petzold & Alan Hindmarsh (1983)
+ *
+ * Copyright (c) 2011 Yu Feng, McWilliam Cosmology Center, Carnegie Mellon University
+ * Copyright (c) 2016 Simon Frost
+ * Copyright (c) 2026 Datagrok
+ *
+ * Licensed under the MIT License. See THIRD_PARTY_LICENSES for the full
+ * license text and provenance chain of the original code.
+ */
+
 import {NordsieckSnapshot} from './common';
 
 /**
