### 2026-02-27 1b41e76
GROK-19735: Diff Grok: Added CVODE
Added the CVODE solver
