+++ src/solver-tools/cvode/cvode_io.ts
@@ -1,0 +1,258 @@
+/**
+ * TypeScript port of SUNDIALS CVODE v7.5.0 (https://github.com/LLNL/sundials)
+ *
+ * Copyright (c) 2025, Lawrence Livermore National Security, University of
+ *   Maryland Baltimore County, and the SUNDIALS contributors.
+ * Copyright (c) 2013-2025, Lawrence Livermore National Security and
+ *   Southern Methodist University.
+ * Copyright (c) 2002-2013, Lawrence Livermore National Security.
+ * Copyright (c) 2026 Datagrok.
+ *
+ * Licensed under the BSD-3-Clause License. See THIRD_PARTY_LICENSES
+ * for the full license text and provenance chain of the original code.
+ */
+
+// CVODE TypeScript port - Configuration and statistics I/O functions
+// Ported from SUNDIALS CVODE v7.5.0 (cvode_io.c)
+
+import {
+  CvodeMem,
+  CV_SUCCESS,
+  CV_SS,
+  CV_SV,
+  ADAMS_Q_MAX,
+  BDF_Q_MAX,
+  CV_ADAMS,
+} from './common';
+
+// ============================================================================
+// Statistics interface
+// ============================================================================
+
+export interface CvodeStats {
+  nSteps: number;
+  nRhsEvals: number;
+  nLinSolvSetups: number;
+  nErrTestFails: number;
+  nNonlinSolvIters: number;
+  nNonlinSolvConvFails: number;
+  nJacEvals: number;
+  nGEvals: number;
+  lastOrder: number;
+  lastStep: number;
+  currentTime: number;
+}
+
+// ============================================================================
+// Tolerance functions
+// ============================================================================
+
+/**
+ * Set scalar relative and absolute tolerances.
+ * Requires rtol >= 0 and atol >= 0.
+ */
+export function cvodeSStolerances(mem: CvodeMem, rtol: number, atol: number): number {
+  if (rtol < 0) {
+    mem.cv_error = 'cvodeSStolerances: rtol must be >= 0';
+    return -1;
+  }
+  if (atol < 0) {
+    mem.cv_error = 'cvodeSStolerances: atol must be >= 0';
+    return -1;
+  }
+
+  mem.cv_reltol = rtol;
+  mem.cv_Sabstol = atol;
+  mem.cv_itol = CV_SS;
+  mem.cv_atolmin0 = (atol === 0);
+
+  return CV_SUCCESS;
+}
+
+/**
+ * Set scalar relative tolerance and vector absolute tolerance.
+ * Requires rtol >= 0 and all atol[i] >= 0.
+ */
+export function cvodeSVtolerances(mem: CvodeMem, rtol: number, atol: Float64Array): number {
+  if (rtol < 0) {
+    mem.cv_error = 'cvodeSVtolerances: rtol must be >= 0';
+    return -1;
+  }
+  for (let i = 0; i < atol.length; i++) {
+    if (atol[i] < 0) {
+      mem.cv_error = `cvodeSVtolerances: atol[${i}] must be >= 0`;
+      return -1;
+    }
+  }
+
+  mem.cv_reltol = rtol;
+  mem.cv_Vabstol = new Float64Array(atol);
+  mem.cv_itol = CV_SV;
+
+  let minAtol = Infinity;
+  for (let i = 0; i < atol.length; i++) {
+    if (atol[i] < minAtol) minAtol = atol[i];
+  }
+  mem.cv_atolmin0 = (minAtol === 0);
+  mem.cv_VabstolMallocDone = true;
+
+  return CV_SUCCESS;
+}
+
+// ============================================================================
+// Optional input functions
[…]
