+++ src/solver-tools/cvode/cvode_adams.ts
@@ -1,0 +1,143 @@
+/**
+ * TypeScript port of SUNDIALS CVODE v7.5.0 (https://github.com/LLNL/sundials)
+ *
+ * Copyright (c) 2025, Lawrence Livermore National Security, University of
+ *   Maryland Baltimore County, and the SUNDIALS contributors.
+ * Copyright (c) 2013-2025, Lawrence Livermore National Security and
+ *   Southern Methodist University.
+ * Copyright (c) 2002-2013, Lawrence Livermore National Security.
+ * Copyright (c) 2026 Datagrok.
+ *
+ * Licensed under the BSD-3-Clause License. See THIRD_PARTY_LICENSES
+ * for the full license text and provenance chain of the original code.
+ */
+
+import { CvodeMem, L_MAX } from './common';
+
+/**
+ * cvAltSum - Alternating sum: sum_{i=0}^{iend} (-1)^i * a[i] / (i + k)
+ */
+function cvAltSum(iend: number, a: Float64Array, k: number): number {
+  if (iend < 0) return 0.0;
+  let sum = 0.0;
+  let sign = 1;
+  for (let i = 0; i <= iend; i++) {
+    sum += sign * (a[i] / (i + k));
+    sign = -sign;
+  }
+  return sum;
+}
+
+/**
+ * cvAdamsStart - Generate product polynomial coefficients for Adams method
+ * Returns hsum.
+ */
+function cvAdamsStart(mem: CvodeMem, m: Float64Array): number {
+  let hsum = mem.cv_h;
+  m[0] = 1.0;
+  for (let i = 1; i <= mem.cv_q; i++) m[i] = 0.0;
+
+  for (let j = 1; j < mem.cv_q; j++) {
+    if (j === mem.cv_q - 1 && mem.cv_qwait === 1) {
+      const sum = cvAltSum(mem.cv_q - 2, m, 2);
+      mem.cv_tq[1] = mem.cv_q * sum / m[mem.cv_q - 2];
+    }
+    const xi_inv = mem.cv_h / hsum;
+    for (let i = j; i >= 1; i--) {
+      m[i] += m[i - 1] * xi_inv;
+    }
+    hsum += mem.cv_tau[j];
+  }
+  return hsum;
+}
+
+/**
+ * cvAdamsFinish - Complete Adams l and tq computation
+ */
+function cvAdamsFinish(mem: CvodeMem, m: Float64Array, M: Float64Array, hsum: number): void {
+  const M0_inv = 1.0 / M[0];
+
+  mem.cv_l[0] = 1.0;
+  for (let i = 1; i <= mem.cv_q; i++) {
+    mem.cv_l[i] = M0_inv * (m[i - 1] / i);
+  }
+
+  const xi = hsum / mem.cv_h;
+  const xi_inv = 1.0 / xi;
+
+  mem.cv_tq[2] = M[1] * M0_inv / xi;
+  mem.cv_tq[5] = xi / mem.cv_l[mem.cv_q];
+
+  if (mem.cv_qwait === 1) {
+    for (let i = mem.cv_q; i >= 1; i--) {
+      m[i] += m[i - 1] * xi_inv;
+    }
+    M[2] = cvAltSum(mem.cv_q, m, 2);
+    mem.cv_tq[3] = M[2] * M0_inv / mem.cv_L;
+  }
+
+  mem.cv_tq[4] = mem.cv_nlscoef / mem.cv_tq[2];
+}
+
+/**
+ * cvSetAdams - Compute Adams method coefficients l[] and tq[]
+ */
[…]
