+++ src/tests/performance.test.ts
@@ -1,12 +1,14 @@
 // Performance tests
 
 import {perfProbs} from '../../index';
-import {implicitMethods, TIMEOUT_MS} from './test-defs';
+import {implicitMethods, perfExclusions, TIMEOUT_MS} from './test-defs';
 
 implicitMethods.forEach((method, name) => {
   describe(`${name}`, () => {
+    const excluded = perfExclusions.get(name);
     perfProbs.forEach((odes) => {
-      test(`${odes.name}`, () => {
+      const testFn = excluded?.has(odes.name) ? test.skip : test;
+      testFn(`${odes.name}`, () => {
         const start = performance.now();
         method(odes);
         const elapsed = performance.now() - start;

[…]
