### 2026-02-27 f06f24a
GROK-19735: Diff Grok: Updated callback actions
Updated callback actions in CVODE and LSODA
