### 2026-02-27 8924248
GROK-19735: Diff Grok: LSODA warmup upd
Updated the warmup stage in LSODA
