### 2026-02-27 de00fbf
GROK-19735: Diff Grok: CVODE warmup upd
Updated the warmup stage in CVODE
