+++ src/solver-tools/cvode/cvode_root.ts
@@ -105,9 +105,9 @@ export function cvRcheck1(mem: CvodeMem): number {
   const nrtfn = mem.cv_nrtfn;
   const N = mem.cv_N;
 
-  for (let i = 0; i < nrtfn; i++) {
+  for (let i = 0; i < nrtfn; i++)
     mem.cv_iroots[i] = 0;
-  }
+
 
   mem.cv_tlo = mem.cv_tn;
   mem.cv_ttol = (Math.abs(mem.cv_tn) + Math.abs(mem.cv_h)) * UROUND * HUNDRED;
@@ -115,9 +115,9 @@ export function cvRcheck1(mem: CvodeMem): number {
   // Evaluate g at initial t and check for zero values
   const retval = mem.cv_gfun!(mem.cv_tlo, mem.cv_zn[0], mem.cv_glo, mem.cv_user_data);
   mem.cv_nge = 1;
-  if (retval !== 0) {
+  if (retval !== 0)
     return CV_RTFUNC_FAIL;
-  }
+
 
   let zroot = false;
   for (let i = 0; i < nrtfn; i++) {
@@ -126,9 +126,9 @@ export function cvRcheck1(mem: CvodeMem): number {
       mem.cv_gactive[i] = 0; // SUNFALSE
     }
   }
-  if (!zroot) {
+  if (!zroot)
     return CV_SUCCESS;
-  }
+
 
   // Some g_i is zero at t0; look at g at t0+(small increment)
   const hratio = Math.max(mem.cv_ttol / Math.abs(mem.cv_h), PT1);
@@ -140,9 +140,9 @@ export function cvRcheck1(mem: CvodeMem): number {
 
   const retval2 = mem.cv_gfun!(tplus, mem.cv_y, mem.cv_ghi, mem.cv_user_data);
   mem.cv_nge++;
-  if (retval2 !== 0) {
+  if (retval2 !== 0)
     return CV_RTFUNC_FAIL;
-  }
+
 
   // Re-activate components where g moved away from zero
   for (let i = 0; i < nrtfn; i++) {
@@ -170,9 +170,9 @@ export function cvRcheck2(mem: CvodeMem): number {
   const nrtfn = mem.cv_nrtfn;
   const N = mem.cv_N;
 
-  if (mem.cv_irfnd === 0) {
+  if (mem.cv_irfnd === 0)
     return CV_SUCCESS;
-  }
+
 
   // Get y at tlo via interpolation
   cvodeGetDky(mem, mem.cv_tlo, 0, mem.cv_y);
@@ -180,14 +180,14 @@ export function cvRcheck2(mem: CvodeMem): number {
   // Evaluate g(tlo, y)
   let retval = mem.cv_gfun!(mem.cv_tlo, mem.cv_y, mem.cv_glo, mem.cv_user_data);
   mem.cv_nge++;
-  if (retval !== 0) {
+  if (retval !== 0)
     return CV_RTFUNC_FAIL;
-  }
+
 
   let zroot = false;
-  for (let i = 0; i < nrtfn; i++) {
+  for (let i = 0; i < nrtfn; i++)
     mem.cv_iroots[i] = 0;
-  }
+
   for (let i = 0; i < nrtfn; i++) {
     if (mem.cv_gactive[i] === 0) continue;
     if (Math.abs(mem.cv_glo[i]) === ZERO) {
@@ -195,9 +195,9 @@ export function cvRcheck2(mem: CvodeMem): number {
       mem.cv_iroots[i] = 1;
     }
   }
-  if (!zroot) {
+  if (!zroot)
     return CV_SUCCESS;
-  }
+
 
[…]
