### 2026-02-27 32ef8f8
GROK-19735: Diff Grok: Fixed LSODA
Updated the LSODA solver
