### 2026-03-02 6c49645
Diff Grok: Added plans
Added plans for CC for porting libs
