+++ CVODE-C-to-TS.md
@@ -1,0 +1,1024 @@
+# CVODE C-to-TypeScript Transformation Plan
+
+This document describes the plan for transforming the CVODE solver (v7.5.0) from
+C to TypeScript, following the patterns established in `LSODA-C-to-TS.md`.
+
+**Source code location:** All TypeScript source files go into `src/cvode/`.
+
+**Test files location:** All tests go into `test/`:
+- `test/cvode.test.ts` — correctness tests (mirrors `test/test.test.ts`)
+- `test/cvode.perf.test.ts` — benchmark/performance tests (mirrors `test/perf.test.ts`)
+
+---
+
+## 1. Scope
+
+### 1.1 What to include
+
+| Feature | Source | Notes |
+|---------|--------|-------|
+| BDF method (stiff) | `cvode.c` | Orders 1-5, primary use case |
+| Adams-Moulton method (non-stiff) | `cvode.c` | Orders 1-12 |
+| Nordsieck history array | `cvode.c` | Core representation, same concept as LSODA |
+| Newton nonlinear solver | `cvode_nls.c` | Default for BDF/stiff |
+| Fixed-point nonlinear solver | `cvode_nls.c` | For Adams/non-stiff |
+| Dense direct linear solver | `cvode_ls.c` + `sunlinsol_dense` | LU with partial pivoting |
+| Difference-quotient Jacobian | `cvode_ls.c` (`cvLsDQJac`) | Default when no user Jacobian |
+| User-supplied Jacobian | `cvode_ls.c` | Optional callback |
+| Error control & step/order selection | `cvode.c` | WRMS norm, adaptive h and q |
+| Rootfinding | `cvode.c` (`cvRcheck*`, `cvRootfind`) | Zero-crossing detection |
+| Dense output | `cvode.c` (`CVodeGetDky`) | Horner evaluation on Nordsieck array |
+| BDF stability limit detection | `cvode.c` (`cvBDFStab`) | Optional safety feature |
+| I/O: tolerances, step limits, stats | `cvode_io.c` (subset) | Configuration and statistics getters |
+
+### 1.2 What to exclude
+
+| Feature | Reason |
+|---------|--------|
+| Band matrix / band solver | Specialized; add later if needed |
+| Sparse matrix / KLU / SuperLU | Specialized; add later if needed |
+| Iterative solvers (SPGMR, SPBCGS, SPTFQMR, SPFGMR) | Krylov methods for large systems; add later if needed |
+| Band preconditioner (`cvode_bandpre.c`) | Requires band solver |
+| BBD preconditioner (`cvode_bbdpre.c`) | Requires band solver + MPI concepts |
+| Diagonal solver (`cvode_diag.c`) | Rarely used; add later if needed |
+| Projection (`cvode_proj.c`) | Constraint projection; add later if needed |
+| Problem resizing (`cvode_resize.c`) | Unusual use case; add later if needed |
+| Constraint enforcement (`cvode_constraints.c`) | Add later if needed |
+| Fused kernel stubs (`cvode_fused_stubs.c`) | Performance optimization not relevant to TS |
+| CLI utilities (`cvode_cli.c`) | Not applicable |
+| Parallel N_Vector variants (OpenMP, CUDA, etc.) | Single-threaded TS runtime |
+| SUNContext / SUNDIALS context | Heavyweight infrastructure; replace with simple config |
+
+### 1.3 Size estimate
+
+| C source (in scope) | Approx lines | TS estimate |
+|----------------------|--------------|-------------|
+| `cvode.c` (core step logic, rootfinding, dense output) | ~4,900 | ~2,500 |
+| `cvode_nls.c` (nonlinear solver) | ~430 | ~250 |
[…]
