### 2026-03-10 #9090 @josemanuel22
@lockwo  Thanks for the reminder, and sorry for the delay. I do not need anything else from you at this stage. I had already gone through the review, but I had not yet created the checklist in the official JOSS format. I’ll generate it now and update it accordingly.
