### 2026-03-13 2caf3fb
Simplify Adams-Bashforth section in README
