### 2026-03-13 ea740f6
Fix typo in README for Adams methods
