### 2026-03-25 #9090 @lockwo
just a heads up on safari the math isn't rendering in the docs
