### 2026-03-25 82334f8
Diff Grok: DOI upd
Update DOI link in README
