### 2026-03-25 #9090 @vmakarichev
Thank you for the review process! Below we address the required post-review author tasks:

✅ **Authors and affiliations**: checked
✅ **Release**: version 1.2.0 - https://github.com/datagrok-ai/diff-grok/releases/tag/v1.2.0
✅ Zenodo: release archived - **DOI: https://doi.org/10.5281/zenodo.19219694**
✅ **Title and author list** in the archive: checked
✅ **License** in the archive: MIT - checked
