### 2026-03-30 #9090 @lockwo
A few minor points on the paper:
- line 61, (Odex-Js , 2025) has a un-needed space before comma (I think this comes from citation)
- Figure 1 seems a bit low resolution? Not sure if there is an easy way to increase DPI since it looks like a screenshot but the other figure is a bit clearer
- Julia should be capitalized on line 143
- R on 155 should also

Also for the archive meta data can you have the version be v1.2.0 (currently it says v2 on zenodo)
