### 2026-03-30 187d05b
Diff Grok: JOSS paper: Updated Fig 1, fixed typos
