+++ paper-joss/paper.bib
@@ -103,7 +103,7 @@
 }
 
 @article{rackauckas2017differentialequations,
-  title={Differentialequations.jl -- a performant and feature-rich ecosystem for solving differential equations in julia},
+  title={Differentialequations.jl -- a performant and feature-rich ecosystem for solving differential equations in {Julia}},
   author={Rackauckas, Christopher and Nie, Qing},
   journal={Journal of Open Research Software},
   volume={5},
@@ -162,7 +162,7 @@
 }
 
 @article{soetaert2010solving,
-  title={Solving Differential Equations in R: Package deSolve},
+  title={Solving Differential Equations in {R}: Package deSolve},
   author={Soetaert, Karline and Petzoldt, Thomas and Setzer, R. Woodrow},
   journal={Journal of Statistical Software},
   volume={33},
@@ -296,7 +296,7 @@ doi = {10.1137/S1064827594276424},
 }
 
 @software{odexjs,
-  title = {odex-js : ODEX in JavaScript},
+  title = {odex-js: ODEX in JavaScript},
   year = {2025},
   url = {https://github.com/littleredcomputer/odex-js}
 }
