### 2026-03-30 #9090 @vmakarichev
Thank you! We have made the following updates:

* In Zenodo, the archive version has been set to 1.2.0.
* Line 61: removed an unnecessary space before a comma.
* Line 143: capitalized “Julia”.
* Line 155: capitalized “R”.
* Figure 1: increased the resolution (DPI).
