### 2026-03-30 356f24a
Diff Grok: Latex basics
Add basic items for the LaTeX features
