### 2026-03-30 a28877f
Diff Grok: LaTeX: Implemented basic LaTeX features
