+++ src/latex-export/transformer/identifier.ts
@@ -1,52 +1,217 @@
-// ============================================================
-// Identifier → LaTeX Transformer
-// ============================================================
-// Converts identifier names to their LaTeX representations.
-//
-// Processing order:
-//   1. Check special constants (PI, Inf)
-//   2. Check Greek alphabet (exact match, then prefix + digit suffix)
-//   3. Single letter + digits → subscript (x1 → x_{1})
-//   4. Multi-letter + trailing digits → mathrm with subscript
-//   5. Chemical formula heuristic (CO2, N2O5)
-//   6. Plain multi-letter → \mathrm{name}
-//
-// See docs/GREEK.md for the complete Greek alphabet mapping.
+/* Identifier-to-LaTeX transformer: converts identifier names to their LaTeX representations. */
 
-/**
- * Convert an identifier to its LaTeX representation.
- *
- * @param name - Identifier string from the expression
- * @returns LaTeX string
+const SPECIAL: Record<string, string> = {
+  'PI': '\\pi',
+  'Inf': '\\infty',
+  'inf': '\\infty',
+};
+
+const GREEK_LOWER: Array<[string, string]> = [
+  ['varepsilon', '\\varepsilon'],
+  ['varsigma', '\\varsigma'],
+  ['vartheta', '\\vartheta'],
+  ['epsilon', '\\epsilon'],
+  ['omicron', 'o'],
+  ['upsilon', '\\upsilon'],
+  ['lambda', '\\lambda'],
+  ['varphi', '\\varphi'],
+  ['varrho', '\\varrho'],
+  ['alpha', '\\alpha'],
+  ['delta', '\\delta'],
+  ['gamma', '\\gamma'],
+  ['kappa', '\\kappa'],
+  ['omega', '\\omega'],
+  ['sigma', '\\sigma'],
+  ['theta', '\\theta'],
+  ['beta', '\\beta'],
+  ['iota', '\\iota'],
+  ['zeta', '\\zeta'],
+  ['chi', '\\chi'],
+  ['eta', '\\eta'],
+  ['phi', '\\phi'],
+  ['psi', '\\psi'],
+  ['rho', '\\rho'],
+  ['tau', '\\tau'],
+  ['mu', '\\mu'],
+  ['nu', '\\nu'],
+  ['pi', '\\pi'],
+  ['xi', '\\xi'],
+];
+
+const GREEK_UPPER: Array<[string, string]> = [
+  ['Epsilon', '\\mathrm{E}'],
+  ['Omicron', '\\mathrm{O}'],
+  ['Upsilon', '\\Upsilon'],
+  ['Lambda', '\\Lambda'],
+  ['Kappa', '\\mathrm{K}'],
+  ['Sigma', '\\Sigma'],
+  ['Theta', '\\Theta'],
+  ['Omega', '\\Omega'],
+  ['Alpha', '\\mathrm{A}'],
+  ['Delta', '\\Delta'],
+  ['Gamma', '\\Gamma'],
+  ['Beta', '\\mathrm{B}'],
+  ['Zeta', '\\mathrm{Z}'],
+  ['Iota', '\\mathrm{I}'],
+  ['Eta', '\\mathrm{H}'],
+  ['Chi', '\\mathrm{X}'],
+  ['Mu', '\\mathrm{M}'],
+  ['Nu', '\\mathrm{N}'],
+  ['Phi', '\\Phi'],
+  ['Psi', '\\Psi'],
+  ['Rho', '\\mathrm{P}'],
+  ['Tau', '\\mathrm{T}'],
+  ['Pi', '\\Pi'],
+  ['Xi', '\\Xi'],
+];
+
+// Combined, sorted by key length descending
+const ALL_GREEK: Array<[string, string]> = [...GREEK_LOWER, ...GREEK_UPPER]
[…]
