### 2026-03-30 e187bb5
Diff Grok: LaTeX: Update examples
