+++ src/latex-export/examples/bioreactor-to-latex-selective.ts
@@ -1,0 +1,147 @@
+/** Example: Bioreactor model → LaTeX with selective sections (no constants, no parameters). */
+
+import {convertIvpToLatex} from '../index';
+
+const MODEL = `#name: Bioreactor
+#description: Bioreactor simulation
+#comment:
+  Source: https://doi.org/10.1074/jbc.RA117.000303.
+#equations:
+
+  d(FFox)/dt = -E11 + E12
+
+  d(KKox)/dt = -E21 + E22
+
+  d(FFred)/dt = E11 - E12 - E31 + E32
+
+  d(KKred)/dt = E21 - E22 - E41 + E42
+
+  d(Ffree)/dt = 2 * (E31 - E32) - E51 + E52
+
+  d(Kfree)/dt = 2 * (E41 - E42) - E51 + E52
+
+  d(FKred)/dt = E51 - E52 - E61 + E62
+
+  d(FKox)/dt = E61 - E62
+
+  d(MEAthiol)/dt = 2 * (-E11 + E12 - E21 + E22 + E31 + E41 - E32 - E42 - E62 - ktox * E71 * E72)
+                   - (MEAthiol + MA) * (Fin + Fper) / VL
+
+  d(CO2)/dt = (Fin * pO2sat * 0.0022 - 2 * Fper * CO2) / VL + OTR - 0.5 * ktox * E71 * E72
+
+  d(yO2P)/dt = -OTR * (VL / (VTV - VL)) * R * T * P + yO2in * qin - yO2P * qout
+
+  d(CYST)/dt = ktox * E71 * E72 - krcyst * CYST * E0 - (Fin + Fper) * CYST / VL
+
+  d(VL)/dt = Fin - Fper
+
+#expressions:
+
+  KF = pow(VL, -0.65) * 0.065 * pow(speed**3 * diam**5 * power / 2.16e12, 0.361)
+
+  MA = MEAthiol * pow(10, pH - pKa2MEA)
+
+  qout = qin - KF * (yO2P * H - CO2) * VL * R * T / (P * 1000)
+
+  OTR = KF * (yO2P * H - CO2)
+
+  Fin = t < switchTime ? 0 : 0.025
+
+  Fper = t < switchTime ? 0.025 : Fin
+
+  E0 = VLinit / VL
+
+  E1 = (MA * E0)**2
+
+  E11 = k1red * FFox * E0 * E1
+
+  E12 = k1ox * FFred * E0
+
+  E31 = k2Fd * FFred * E0
+
+  E32 = k2Fa * (Ffree * E0)**2 * E1
+
+  E21 = k1red * KKox * E0 * E1
+
+  E22 = k1ox * KKred * E0
+
+  E41 = k2Kd * KKred * E0
+
+  E42 = k2Ka * (Kfree * E0)**2 * E1
+
+  E51 = k3FKa * Ffree * E0 * Kfree * E0
+
+  E52 = k3FKd * FKred * E0
+
+  E61 = k4ox * FKred * E0 * (CYST * E0)**2
+
+  E62 = k4red * FKox * E0 * E1
+
+  E70 = E0 * CO2
+
+  E71 = (MEAthiol * E0)**2
+
+  E72 = (E70 >= 0) ? sqrt(E70) : 0
+
+#argument: t
+  t0 = 0.0   {units: min; caption: Initial; category: Time}
+  t1 = 1000  {units: min; caption: Final;   category: Time; min: 500; max: 1000}
+   h = 1     {units: min; caption: Step;    category: Time; min: 0.1; max: 2}
+
+#inits:
+  FFox     = 0.2   {units: mmol/L; category: Initial values; min: 0.15; max: 0.25; step: 0.01}
[…]
