### 2026-03-31 e9ec90c
Diff Grok: LaTeX: Loop upd
Update export to md and latex of models with loops
