+++ src/latex-export/generator/document-builder.ts
@@ -57,6 +57,19 @@ export function buildLatexDocument(
     parts.push(`Before each cycle:\n\\begin{align}\n${lines.join(' \\\\\n')}\n\\end{align}`);
   }
 
+  const stages = extractStageInfos(model);
+  if (stages.length > 0) {
+    parts.push('\\subsection{Stage Transitions}');
+    for (const stage of stages) {
+      const dur = stage.duration ? ` (duration: $${expressionToLatex(stage.duration)}$)` : '';
+      parts.push(`\\textbf{${stage.stageName}}${dur}`);
+      if (stage.transitions.length > 0) {
+        const lines = stage.transitions.map((t) => `  ${renderTransitionLine(t)}`);
+        parts.push(`Before this stage:\n\\begin{align}\n${lines.join(' \\\\\n')}\n\\end{align}`);
+      }
+    }
+  }
+
   if (opts.includeInits && model.inits.length > 0) {
     parts.push(`\\subsection{${initsTitle(model)}}`);
     parts.push(buildLatexTable(model.inits));
@@ -118,6 +131,19 @@ export function buildMarkdownDocument(
     parts.push(`Before each cycle:\n\n$$\n\\begin{aligned}\n${lines.join(' \\\\\n')}\n\\end{aligned}\n$$`);
   }
 
+  const stages = extractStageInfos(model);
+  if (stages.length > 0) {
+    parts.push('### Stage Transitions');
+    for (const stage of stages) {
+      const dur = stage.duration ? ` (duration: $${expressionToLatex(stage.duration)}$)` : '';
+      parts.push(`**${stage.stageName}**${dur}`);
+      if (stage.transitions.length > 0) {
+        const lines = stage.transitions.map((t) => `  ${renderTransitionLine(t)}`);
+        parts.push(`Before this stage:\n\n$$\n\\begin{aligned}\n${lines.join(' \\\\\n')}\n\\end{aligned}\n$$`);
+      }
+    }
+  }
+
   if (opts.includeInits && model.inits.length > 0) {
     parts.push(`### ${initsTitle(model)}`);
     parts.push(buildMarkdownTable(model.inits));
@@ -149,8 +175,12 @@ function buildArgRangeLatex(model: ParsedModel): string {
   const step = model.argument.entries[2].value;
 
   const loopInfo = extractLoopInfo(model);
+  const stages = extractStageInfos(model);
   let effectiveFinish = finish;
+  let suffix = '';
+
   if (loopInfo) {
+    // Cyclic model: total = start + count * (finish - start)
     const s = parseFloat(start);
     const f = parseFloat(finish);
     const n = parseFloat(loopInfo.count);
@@ -158,15 +188,21 @@ function buildArgRangeLatex(model: ParsedModel): string {
       effectiveFinish = String(s + n * (f - s));
     else
       effectiveFinish = `${start} + ${loopInfo.count} \\cdot (${finish} - ${start})`;
-  }
-
-  let result = `${arg} \\in \\left[${start},\\, ${effectiveFinish}\\right], \\quad \\Delta ${arg} = ${step}`;
-  if (loopInfo) {
     const duration = (!isNaN(parseFloat(start)) && !isNaN(parseFloat(finish))) ?
       String(parseFloat(finish) - parseFloat(start)) : `${finish} - ${start}`;
-    result += `, \\quad ${arg}_{\\text{cycle}} = ${duration}`;
+    suffix = `, \\quad ${arg}_{\\text{cycle}} = ${duration}`;
+  } else if (stages.length > 0) {
+    // Multistage model: total = start + stage1_duration + stage2_duration + ...
+    const durations = [finish, ...stages.map((s) => s.duration).filter(Boolean) as string[]];
[…]
