### 2026-04-01 #9090 @lockwo
@vmakarichev - please let me know if you see any issues with the proof that this generates
