### 2026-04-01 e2395b0
Diff Grok: LaTeX: Readme update
