### 2026-04-01 a645175
minor changes in paper.md
