+++ paper-joss/paper.md
@@ -46,12 +46,13 @@ While a broad ecosystem of mature numerical solvers exists, effective **interact
 Diff Studio is a **browser-native environment for solving and exploring IVPs**, designed to support interactive, collaborative, and reproducible ODE modeling without custom software development or local environment setup. It enables users to define models declaratively, explore their behavior interactively, and share fully reproducible simulations through the browser.
 
 Diff Studio consists of two components:
-1. **Diff Grok**:  general-purpose high-performance TypeScript library for solving ODE systems defined in a declarative form
-2. **Diff Studio**: web application that integrates Diff Grok into the **Datagrok** scientific computing platform. It provides autogenerated user interface, rich interactive visualizations, and the ability to manage ODE models and simulations.
+
+1. **Diff Grok** is a general-purpose high-performance TypeScript library for solving ODE systems defined in a declarative form.
+2. **Diff Studio** is a web application that integrates Diff Grok into the **Datagrok** scientific computing platform. It provides autogenerated user interface, rich interactive visualizations, and the ability to manage ODE models and simulations.
 
 # Statement of need
 
-ODEs can be solved either analytically or numerically. Analytic methods provide exact solutions but apply only to limited classes of problems and are often impractical due to their complexity [@hairer2008solving1]. Numerical methods, which compute approximate solutions, are therefore the dominant approach. A wide range of such methods has been developed [@hairer2008solving1; @hairer2002solving2] and is available in established scientific computing tools, including SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials], Julia DifferentialEquations [@rackauckas2017differentialequations], SciPy [@2020SciPyNMeth], Maple [@maple2025], Mathematica [@Mathematica2024], Matlab [@MATLAB], and deSolve [@soetaert2010solving].
+ODEs can be solved either analytically or numerically. Analytic methods provide exact solutions but apply only to limited classes of problems and are often impractical due to their complexity [@hairer2008solving1]. Numerical methods, which compute approximate solutions, are therefore the dominant approach. A wide range of such methods have been developed [@hairer2008solving1; @hairer2002solving2] and are available in established scientific computing tools, including SUNDIALS [@gardner2022sundials; @hindmarsh2005sundials], Julia DifferentialEquations [@rackauckas2017differentialequations], SciPy [@2020SciPyNMeth], Maple [@maple2025], Mathematica [@Mathematica2024], Matlab [@MATLAB], and deSolve [@soetaert2010solving].
 
 These tools provide robust and efficient solvers but are primarily designed for **desktop or scripting-centric workflows**. In practice, scientific modeling is often exploratory: researchers iteratively adjust parameters, inspect transient behavior, compare scenarios, and refine model structure. Collaboration across teams and institutions is also common. Existing solutions typically require local installation, environment configuration, and custom scripting, which complicates sharing models and reproducing results. Building interactive analysis tools on top of these systems often requires substantial additional programming effort, shifting focus away from scientific inquiry toward software engineering.
 
