### 2026-04-01 #9090 @danielskatz
👋 @vmakarichev - as track editor, I've proofread the paper, and suggest the changes in https://github.com/datagrok-ai/diff-grok/pull/18 - please merge this, or let me know what you disagree with.
