### 2026-04-01 389c778
Add citation section to README
Added citation section with reference article for research purposes.
