### 2026-04-01 #9090 @danielskatz
Congratulations to @vmakarichev (Victor Makarichev) and co-authors on your JOSS publication!!

JOSS fully depends on volunteers - we couldn't do this without you
